Get bangumi calendar

Get bangumi calendar

bgm_calendar(format = c("list", "table", "default"))

Arguments

format
Result will be a list of json data when set to list (default), response() object when set to default. Set to table will convert the list to a data.frame for ease of use.

Value

A list of response content or a response() object according to the format option.

Examples

bgm_calendar("table")

#>         id                          url type
#>  1: 293193 http://bgm.tv/subject/293193    2
#>  2: 297168 http://bgm.tv/subject/297168    2
#>  3: 297344 http://bgm.tv/subject/297344    2
#>  4: 301601 http://bgm.tv/subject/301601    2
#>  5: 315574 http://bgm.tv/subject/315574    2
#>  6: 320824 http://bgm.tv/subject/320824    2
#>  7: 195723 http://bgm.tv/subject/195723    2
#>  8: 278031 http://bgm.tv/subject/278031    2
#>  9: 308961 http://bgm.tv/subject/308961    2
#> 10: 310950 http://bgm.tv/subject/310950    2
#> 11: 316607 http://bgm.tv/subject/316607    2
#> 12: 278998 http://bgm.tv/subject/278998    2
#> 13: 280615 http://bgm.tv/subject/280615    2
#> 14: 295017 http://bgm.tv/subject/295017    2
#> 15: 297224 http://bgm.tv/subject/297224    2
#> 16: 297370 http://bgm.tv/subject/297370    2
#> 17: 298698 http://bgm.tv/subject/298698    2
#> 18: 302639 http://bgm.tv/subject/302639    2
#> 19: 303172 http://bgm.tv/subject/303172    2
#> 20: 312700 http://bgm.tv/subject/312700    2
#> 21: 316247 http://bgm.tv/subject/316247    2
#> 22: 262897 http://bgm.tv/subject/262897    2
#> 23: 279202 http://bgm.tv/subject/279202    2
#> 24: 282000 http://bgm.tv/subject/282000    2
#> 25: 293945 http://bgm.tv/subject/293945    2
#> 26: 296261 http://bgm.tv/subject/296261    2
#> 27: 302802 http://bgm.tv/subject/302802    2
#> 28: 304827 http://bgm.tv/subject/304827    2
#> 29: 252782 http://bgm.tv/subject/252782    2
#> 30: 292527 http://bgm.tv/subject/292527    2
#> 31: 297396 http://bgm.tv/subject/297396    2
#> 32: 302797 http://bgm.tv/subject/302797    2
#> 33: 306147 http://bgm.tv/subject/306147    2
#> 34: 309002 http://bgm.tv/subject/309002    2
#> 35: 314488 http://bgm.tv/subject/314488    2
#> 36: 316104 http://bgm.tv/subject/316104    2
#> 37: 316795 http://bgm.tv/subject/316795    2
#> 38: 322926 http://bgm.tv/subject/322926    2
#> 39:    899    http://bgm.tv/subject/899    2
#> 40: 278815 http://bgm.tv/subject/278815    2
#> 41: 296875 http://bgm.tv/subject/296875    2
#> 42: 304065 http://bgm.tv/subject/304065    2
#> 43: 314447 http://bgm.tv/subject/314447    2
#> 44: 315069 http://bgm.tv/subject/315069    2
#> 45: 315456 http://bgm.tv/subject/315456    2
#> 46:    975    http://bgm.tv/subject/975    2
#> 47: 269565 http://bgm.tv/subject/269565    2
#> 48: 277554 http://bgm.tv/subject/277554    2
#> 49: 282684 http://bgm.tv/subject/282684    2
#> 50: 295766 http://bgm.tv/subject/295766    2
#> 51: 296151 http://bgm.tv/subject/296151    2
#> 52: 296941 http://bgm.tv/subject/296941    2
#> 53: 315804 http://bgm.tv/subject/315804    2
#> 54: 319894 http://bgm.tv/subject/319894    2
#> 55: 321092 http://bgm.tv/subject/321092    2
#> 56: 321289 http://bgm.tv/subject/321289    2
#>         id                          url type
#>                                                                 name
#>  1: たとえばラストダンジョン前の村の少年が序盤の街で暮らすような物語
#>  2:                                アズールレーン びそくぜんしんっ！
#>  3:                                              真・中華一番！ 続編
#>  4:                                                 裏世界ピクニック
#>  5:                               ウマ娘 プリティーダービー Season 2
#>  6:                                     WAVE!!～サーフィンやっぺ!!～
#>  7:                                                         ゲキドル
#>  8:                                 転生したらスライムだった件 第2期
#>  9:                                      文豪ストレイドッグス わん！
#> 10:                              ワールドウィッチーズ 発進しますっ！
#> 11:                                   ワンダーエッグ・プライオリティ
#> 12:                                                       装甲娘戦機
#> 13:                                                      アイ★チュウ
#> 14:                                               回復術士のやり直し
#> 15:                                              BEASTARS 2nd Season
#> 16:                                             オルタンシア・サーガ
#> 17:                                        ログ・ホライズン 円卓崩壊
#> 18:                                            七つの大罪 憤怒の審判
#> 19:                            魔術士オーフェンはぐれ旅 キムラック編
#> 20:                                                 伍六七之玄武国篇
#> 21:                Re:ゼロから始める異世界生活 2nd season 後半クール
#> 22:                                             ゆるキャン△ SEASON 2
#> 23:                                      約束のネバーランド Season 2
#> 24:                                                    五等分の花嫁∬
#> 25:                                        2.43 清陰高校男子バレー部
#> 26:                                              Dr.STONE STONE WARS
#> 27:                                           SHOW BY ROCK!! STARS!!
#> 28:                                               天地創造デザイン部
#> 29:                                              蜘蛛ですが、なにか?
#> 30:                                                 弱キャラ友崎くん
#> 31:                                                   バック・アロウ
#> 32:                                               WIXOSS DIVA(A)LIVE
#> 33:                                       俺だけ入れる隠しダンジョン
#> 34:                                              銀魂 THE SEMI FINAL
#> 35:                                                     プレイタの傷
#> 36:                                                   アイドールズ！
#> 37:                                                         幼女社長
#> 38:                                             おとなの防具屋さんII
#> 39:                                                     名探偵コナン
#> 40:                                                   はたらく細胞!!
#> 41:                                     ワールドトリガー 2ndシーズン
#> 42:                                                はたらく細胞BLACK
#> 43:                                                     怪病医ラムネ
#> 44:                                                         ホリミヤ
#> 45:                                               SK∞ エスケーエイト
#> 46:                                                        ONE PIECE
#> 47:                                              EX-ARM エクスアーム
#> 48:                              無職転生 ～異世界行ったら本気だす～
#> 49:                                      のんのんびより のんすとっぷ
#> 50:                                    スケートリーディング☆スターズ
#> 51:                                                      IDOLY PRIDE
#> 52:                                                         怪物事変
#> 53:                               八十亀ちゃんかんさつにっき 3さつめ
#> 54:                     じみへんっ!!～地味子を変えちゃう純異性交遊～
#> 55:                                                     闇芝居 第8期
#> 56:                                              アイカツプラネット!
#>                                                                 name
#>                                          name_cn summary   air_date air_weekday
#>  1: 好比是最终迷宫前的少年到新手村生活一般的故事         2021-01-04           1
#>  2:                          碧蓝航线 微速前进！         2021-01-11           1
#>  3:                              厨神小当家 续篇         2021-01-11           1
#>  4:                                   里世界郊游         2021-01-04           1
#>  5:                   赛马娘 Pretty Derby 第二季         2021-01-04           1
#>  6:                       来冲浪吧！！美少年！！         2021-01-11           1
#>  7:                                       剧偶像         2021-01-05           2
#>  8:            关于我转生变成史莱姆这档事 第二季         2021-01-12           2
#>  9:                                文豪野犬 汪！         2021-01-12           2
#> 10:                              世界魔女 起飞！         2021-01-12           2
#> 11:                                     奇蛋物语         2021-01-12           2
#> 12:                                   装甲娘战机         2021-01-06           3
#> 13:                               ICHU偶像进行曲         2021-01-06           3
#> 14:                           回复术士的重来人生         2021-01-13           3
#> 15:                                                      2021-01-06           3
#> 16:                                   苍之骑士团         2021-01-06           3
#> 17:                        记录的地平线 圆桌崩坏         2021-01-13           3
#> 18:                            七大罪 愤怒的审判         2021-01-06           3
#> 19:              魔术士奥芬的无赖之旅 基姆拉克篇         2021-01-20           3
#> 20:                                                      2021-01-27           3
#> 21:     Re：从零开始的异世界生活 第二季 后半部分         2021-01-06           3
#> 22:                             摇曳露营△ 第二季         2021-01-07           4
#> 23:                          约定的梦幻岛 第二季         2021-01-07           4
#> 24:                                五等分的新娘∬         2021-01-07           4
#> 25:                      2.43 清阴高校男子排球部         2021-01-07           4
#> 26:                            新石纪 STONE WARS         2021-01-14           4
#> 27:                       SHOW BY ROCK!! STARS!!         2021-01-07           4
#> 28:                               天地创造设计部         2021-01-07           4
#> 29:                           转生成蜘蛛又怎样！         2021-01-08           5
#> 30:                                 弱角友崎同学         2021-01-08           5
#> 31:                                   BACK ARROW         2021-01-08           5
#> 32:                                                      2021-01-08           5
#> 33:                       只有我进入的隐藏地下城         2021-01-08           5
#> 34:                          银魂 THE SEMI FINAL         2021-01-15           5
#> 35:                                  Praeter之伤         2021-01-08           5
#> 36:                                                      2021-01-08           5
#> 37:                                     幼女社长         2021-01-01           5
#> 38:                          大人的防具店 第二季         2021-01-08           5
#> 39:                                   名侦探柯南         1996-01-08           6
#> 40:                                 工作细胞！！         2021-01-09           6
#> 41:                            境界触发者 第二季         2021-01-09           6
#> 42:                                工作细胞BLACK         2021-01-09           6
#> 43:                                 怪病医拉姆尼         2021-01-09           6
#> 44:                                     堀与宫村         2021-01-09           6
#> 45:                                     无限滑板         2021-01-09           6
#> 46:                                       海贼王         1999-10-20           7
#> 47:                                       EX-ARM         2021-01-10           7
#> 48:            无职转生 ～在异世界认真地活下去～         2021-01-10           7
#> 49:                         悠哉日常大王 Nonstop         2021-01-10           7
#> 50:                                花样滑冰stars         2021-01-10           7
#> 51:                                  IDOLY PRIDE         2021-01-10           7
#> 52:                                     怪物事变         2021-01-10           7
#> 53:                      八十龟酱观察日记 第三册         2021-01-10           7
#> 54:         地味变!!～改变土妹子的纯洁异性交往～         2021-01-03           7
#> 55:                                暗芝居 第八季         2021-01-10           7
#> 56:                             偶像活动Planet！         2021-01-10           7
#>                                          name_cn summary   air_date air_weekday
#>     rating.total rating.count.10 rating.count.9 rating.count.8 rating.count.7
#>  1:           68               3              0              2             30
#>  2:           29               2              0              2             12
#>  3:           NA              NA             NA             NA             NA
#>  4:           86               3              0             14             37
#>  5:           60               0              5             28             23
#>  6:            1               0              0              0              1
#>  7:           29               0              0              2              8
#>  8:           58               8              1             16             20
#>  9:           NA              NA             NA             NA             NA
#> 10:            4               0              0              2              2
#> 11:            2               0              0              0              2
#> 12:           17               0              0              0              2
#> 13:            5               0              0              0              2
#> 14:            5               1              0              0              1
#> 15:           96               1              3             60             25
#> 16:           25               1              0              0              4
#> 17:           11               5              0              0              5
#> 18:            4               0              0              3              0
#> 19:           NA              NA             NA             NA             NA
#> 20:            3               0              0              1              1
#> 21:          150              16             14             67             30
#> 22:          263              65             99             72             19
#> 23:          124               3              5             58             39
#> 24:           76               7              0             12             38
#> 25:           20               1              0              2              7
#> 26:           16               1              0             10              4
#> 27:            7               0              0              3              4
#> 28:           56               2              1              9             20
#> 29:           78               3              0              8             35
#> 30:           90               3              1              2             29
#> 31:           33               2              0              3             18
#> 32:           10               0              0              0              3
#> 33:           32               2              1              3             12
#> 34:            2               0              0              0              0
#> 35:            8               0              0              0              3
#> 36:           NA              NA             NA             NA             NA
#> 37:          127               0              0              3             36
#> 38:            7               1              0              1              1
#> 39:         4610             542            536           1369           1439
#> 40:           62               5              2             17             27
#> 41:            2               0              1              0              1
#> 42:           68               7              4             34             18
#> 43:            2               0              0              1              0
#> 44:           76               2              3             33             22
#> 45:           17               1              4              8              4
#> 46:         6795            2333           1433           1507            843
#> 47:            1               0              0              0              0
#> 48:           43              12              6             16              6
#> 49:           36               7              7             15              4
#> 50:           23               0              0              4              7
#> 51:            2               0              0              0              2
#> 52:            5               0              0              1              3
#> 53:           NA              NA             NA             NA             NA
#> 54:            7               1              0              3              1
#> 55:            1               0              0              0              0
#> 56:           NA              NA             NA             NA             NA
#>     rating.total rating.count.10 rating.count.9 rating.count.8 rating.count.7
#>     rating.count.6 rating.count.5 rating.count.4 rating.count.3 rating.count.2
#>  1:             23              5              1              0              1
#>  2:              8              3              2              0              0
#>  3:             NA             NA             NA             NA             NA
#>  4:             26              4              0              1              0
#>  5:              3              1              0              0              0
#>  6:              0              0              0              0              0
#>  7:             15              2              2              0              0
#>  8:             10              1              1              0              0
#>  9:             NA             NA             NA             NA             NA
#> 10:              0              0              0              0              0
#> 11:              0              0              0              0              0
#> 12:              7              5              2              1              0
#> 13:              1              0              2              0              0
#> 14:              1              0              0              0              0
#> 15:              4              1              1              0              0
#> 16:              5              7              5              2              1
#> 17:              0              0              0              0              0
#> 18:              1              0              0              0              0
#> 19:             NA             NA             NA             NA             NA
#> 20:              0              0              0              0              0
#> 21:             16              3              1              0              1
#> 22:              2              2              0              0              0
#> 23:             11              4              1              0              0
#> 24:             14              3              1              1              0
#> 25:              8              1              0              0              0
#> 26:              1              0              0              0              0
#> 27:              0              0              0              0              0
#> 28:             15              8              1              0              0
#> 29:             21              9              0              1              0
#> 30:             24             12              7              2              2
#> 31:              7              3              0              0              0
#> 32:              5              2              0              0              0
#> 33:              5              3              2              1              0
#> 34:              2              0              0              0              0
#> 35:              0              2              2              1              0
#> 36:             NA             NA             NA             NA             NA
#> 37:             62             20              5              0              0
#> 38:              1              2              0              1              0
#> 39:            551            131             22              5              2
#> 40:              7              2              0              0              1
#> 41:              0              0              0              0              0
#> 42:              4              1              0              0              0
#> 43:              1              0              0              0              0
#> 44:             13              2              1              0              0
#> 45:              0              0              0              0              0
#> 46:            370            161             44             19             27
#> 47:              0              0              0              1              0
#> 48:              0              1              0              0              0
#> 49:              3              0              0              0              0
#> 50:              3              6              1              1              0
#> 51:              0              0              0              0              0
#> 52:              1              0              0              0              0
#> 53:             NA             NA             NA             NA             NA
#> 54:              1              0              1              0              0
#> 55:              0              1              0              0              0
#> 56:             NA             NA             NA             NA             NA
#>     rating.count.6 rating.count.5 rating.count.4 rating.count.3 rating.count.2
#>     rating.count.1 rating.score rank
#>  1:              3          6.3 3533
#>  2:              0          6.6   NA
#>  3:             NA           NA   NA
#>  4:              1          6.8 2571
#>  5:              0          7.6 1101
#>  6:              0          7.0   NA
#>  7:              0          6.2   NA
#>  8:              1          7.4 1365
#>  9:             NA           NA   NA
#> 10:              0          7.5   NA
#> 11:              0          7.0   NA
#> 12:              0          5.4   NA
#> 13:              0          5.6   NA
#> 14:              2          5.0   NA
#> 15:              1          7.6  967
#> 16:              0          5.2   NA
#> 17:              1          7.8   NA
#> 18:              0          7.5   NA
#> 19:             NA           NA   NA
#> 20:              1          5.3   NA
#> 21:              2          7.7  688
#> 22:              4          8.7   48
#> 23:              3          7.3 1331
#> 24:              0          7.1 1821
#> 25:              1          6.5   NA
#> 26:              0          7.8   NA
#> 27:              0          7.4   NA
#> 28:              0          6.7 2533
#> 29:              1          6.6 2800
#> 30:              8          5.6 4892
#> 31:              0          6.9   NA
#> 32:              0          6.1   NA
#> 33:              3          6.1   NA
#> 34:              0          6.0   NA
#> 35:              0          5.3   NA
#> 36:             NA           NA   NA
#> 37:              1          6.1 4524
#> 38:              0          6.3   NA
#> 39:             13          7.7  525
#> 40:              1          7.2   NA
#> 41:              0          8.0   NA
#> 42:              0          7.8   NA
#> 43:              0          7.0   NA
#> 44:              0          7.3 1366
#> 45:              0          8.1   NA
#> 46:             58          8.5   43
#> 47:              0          3.0   NA
#> 48:              2          8.2   NA
#> 49:              0          8.3   NA
#> 50:              1          6.0   NA
#> 51:              0          7.0   NA
#> 52:              0          7.0   NA
#> 53:             NA           NA   NA
#> 54:              0          7.3   NA
#> 55:              0          5.0   NA
#> 56:             NA           NA   NA
#>     rating.count.1 rating.score rank
#>                                              images.large
#>  1: http://lain.bgm.tv/pic/cover/l/98/e5/293193_d88nd.jpg
#>  2: http://lain.bgm.tv/pic/cover/l/42/fd/297168_hx33I.jpg
#>  3: http://lain.bgm.tv/pic/cover/l/e4/dc/297344_qhoyE.jpg
#>  4: http://lain.bgm.tv/pic/cover/l/e6/f6/301601_9TO1X.jpg
#>  5: http://lain.bgm.tv/pic/cover/l/1c/6b/315574_tLgrT.jpg
#>  6: http://lain.bgm.tv/pic/cover/l/1a/8b/320824_hQG00.jpg
#>  7: http://lain.bgm.tv/pic/cover/l/91/4d/195723_z322N.jpg
#>  8: http://lain.bgm.tv/pic/cover/l/e6/78/278031_202kC.jpg
#>  9: http://lain.bgm.tv/pic/cover/l/09/50/308961_2xti9.jpg
#> 10: http://lain.bgm.tv/pic/cover/l/56/d9/310950_5mmvt.jpg
#> 11: http://lain.bgm.tv/pic/cover/l/dd/60/316607_Qukn6.jpg
#> 12: http://lain.bgm.tv/pic/cover/l/2d/0d/278998_WRj0d.jpg
#> 13: http://lain.bgm.tv/pic/cover/l/c0/76/280615_zfF1K.jpg
#> 14: http://lain.bgm.tv/pic/cover/l/c7/3f/295017_EXE8i.jpg
#> 15: http://lain.bgm.tv/pic/cover/l/f4/44/297224_eE7lv.jpg
#> 16: http://lain.bgm.tv/pic/cover/l/d1/43/297370_Eue45.jpg
#> 17: http://lain.bgm.tv/pic/cover/l/16/c6/298698_T13k1.jpg
#> 18: http://lain.bgm.tv/pic/cover/l/fe/e5/302639_AejA9.jpg
#> 19: http://lain.bgm.tv/pic/cover/l/02/23/303172_L1cyr.jpg
#> 20: http://lain.bgm.tv/pic/cover/l/5d/db/312700_vpftp.jpg
#> 21: http://lain.bgm.tv/pic/cover/l/aa/64/316247_a6jUR.jpg
#> 22: http://lain.bgm.tv/pic/cover/l/0f/50/262897_Lfm1T.jpg
#> 23: http://lain.bgm.tv/pic/cover/l/64/81/279202_tMc9C.jpg
#> 24: http://lain.bgm.tv/pic/cover/l/6c/00/282000_Hax2m.jpg
#> 25: http://lain.bgm.tv/pic/cover/l/65/f5/293945_89GwC.jpg
#> 26: http://lain.bgm.tv/pic/cover/l/50/b5/296261_i3iaJ.jpg
#> 27: http://lain.bgm.tv/pic/cover/l/51/8d/302802_Df3oo.jpg
#> 28: http://lain.bgm.tv/pic/cover/l/b6/39/304827_Bbn7q.jpg
#> 29: http://lain.bgm.tv/pic/cover/l/83/73/252782_6PjZ1.jpg
#> 30: http://lain.bgm.tv/pic/cover/l/c8/42/292527_YtaWi.jpg
#> 31: http://lain.bgm.tv/pic/cover/l/94/8c/297396_xJK3Q.jpg
#> 32: http://lain.bgm.tv/pic/cover/l/f2/c1/302797_SZfJM.jpg
#> 33: http://lain.bgm.tv/pic/cover/l/fc/73/306147_o4qLV.jpg
#> 34: http://lain.bgm.tv/pic/cover/l/27/3c/309002_BQIXI.jpg
#> 35: http://lain.bgm.tv/pic/cover/l/1d/86/314488_tnvn3.jpg
#> 36: http://lain.bgm.tv/pic/cover/l/b3/c9/316104_1yKUQ.jpg
#> 37: http://lain.bgm.tv/pic/cover/l/7b/ef/316795_YxeeC.jpg
#> 38: http://lain.bgm.tv/pic/cover/l/0c/80/322926_cwAAw.jpg
#> 39:    http://lain.bgm.tv/pic/cover/l/01/88/899_Q3F3X.jpg
#> 40: http://lain.bgm.tv/pic/cover/l/ae/11/278815_veMhu.jpg
#> 41: http://lain.bgm.tv/pic/cover/l/97/02/296875_3Z9k9.jpg
#> 42: http://lain.bgm.tv/pic/cover/l/bf/76/304065_4Crea.jpg
#> 43: http://lain.bgm.tv/pic/cover/l/a9/72/314447_5CQ03.jpg
#> 44: http://lain.bgm.tv/pic/cover/l/23/a2/315069_QlT2S.jpg
#> 45: http://lain.bgm.tv/pic/cover/l/8b/0f/315456_Q6m1s.jpg
#> 46:    http://lain.bgm.tv/pic/cover/l/92/97/975_YKuWd.jpg
#> 47: http://lain.bgm.tv/pic/cover/l/18/80/269565_5W5I1.jpg
#> 48: http://lain.bgm.tv/pic/cover/l/8b/00/277554_tG28z.jpg
#> 49: http://lain.bgm.tv/pic/cover/l/ac/b7/282684_0B301.jpg
#> 50: http://lain.bgm.tv/pic/cover/l/3b/af/295766_dJ65o.jpg
#> 51: http://lain.bgm.tv/pic/cover/l/14/f6/296151_b8ZGB.jpg
#> 52: http://lain.bgm.tv/pic/cover/l/6d/7c/296941_j4dd2.jpg
#> 53: http://lain.bgm.tv/pic/cover/l/86/05/315804_GPogW.jpg
#> 54: http://lain.bgm.tv/pic/cover/l/ff/d0/319894_9GIo6.jpg
#> 55: http://lain.bgm.tv/pic/cover/l/29/51/321092_yFId5.jpg
#> 56: http://lain.bgm.tv/pic/cover/l/2f/e7/321289_1M1GJ.jpg
#>                                              images.large
#>                                             images.common
#>  1: http://lain.bgm.tv/pic/cover/c/98/e5/293193_d88nd.jpg
#>  2: http://lain.bgm.tv/pic/cover/c/42/fd/297168_hx33I.jpg
#>  3: http://lain.bgm.tv/pic/cover/c/e4/dc/297344_qhoyE.jpg
#>  4: http://lain.bgm.tv/pic/cover/c/e6/f6/301601_9TO1X.jpg
#>  5: http://lain.bgm.tv/pic/cover/c/1c/6b/315574_tLgrT.jpg
#>  6: http://lain.bgm.tv/pic/cover/c/1a/8b/320824_hQG00.jpg
#>  7: http://lain.bgm.tv/pic/cover/c/91/4d/195723_z322N.jpg
#>  8: http://lain.bgm.tv/pic/cover/c/e6/78/278031_202kC.jpg
#>  9: http://lain.bgm.tv/pic/cover/c/09/50/308961_2xti9.jpg
#> 10: http://lain.bgm.tv/pic/cover/c/56/d9/310950_5mmvt.jpg
#> 11: http://lain.bgm.tv/pic/cover/c/dd/60/316607_Qukn6.jpg
#> 12: http://lain.bgm.tv/pic/cover/c/2d/0d/278998_WRj0d.jpg
#> 13: http://lain.bgm.tv/pic/cover/c/c0/76/280615_zfF1K.jpg
#> 14: http://lain.bgm.tv/pic/cover/c/c7/3f/295017_EXE8i.jpg
#> 15: http://lain.bgm.tv/pic/cover/c/f4/44/297224_eE7lv.jpg
#> 16: http://lain.bgm.tv/pic/cover/c/d1/43/297370_Eue45.jpg
#> 17: http://lain.bgm.tv/pic/cover/c/16/c6/298698_T13k1.jpg
#> 18: http://lain.bgm.tv/pic/cover/c/fe/e5/302639_AejA9.jpg
#> 19: http://lain.bgm.tv/pic/cover/c/02/23/303172_L1cyr.jpg
#> 20: http://lain.bgm.tv/pic/cover/c/5d/db/312700_vpftp.jpg
#> 21: http://lain.bgm.tv/pic/cover/c/aa/64/316247_a6jUR.jpg
#> 22: http://lain.bgm.tv/pic/cover/c/0f/50/262897_Lfm1T.jpg
#> 23: http://lain.bgm.tv/pic/cover/c/64/81/279202_tMc9C.jpg
#> 24: http://lain.bgm.tv/pic/cover/c/6c/00/282000_Hax2m.jpg
#> 25: http://lain.bgm.tv/pic/cover/c/65/f5/293945_89GwC.jpg
#> 26: http://lain.bgm.tv/pic/cover/c/50/b5/296261_i3iaJ.jpg
#> 27: http://lain.bgm.tv/pic/cover/c/51/8d/302802_Df3oo.jpg
#> 28: http://lain.bgm.tv/pic/cover/c/b6/39/304827_Bbn7q.jpg
#> 29: http://lain.bgm.tv/pic/cover/c/83/73/252782_6PjZ1.jpg
#> 30: http://lain.bgm.tv/pic/cover/c/c8/42/292527_YtaWi.jpg
#> 31: http://lain.bgm.tv/pic/cover/c/94/8c/297396_xJK3Q.jpg
#> 32: http://lain.bgm.tv/pic/cover/c/f2/c1/302797_SZfJM.jpg
#> 33: http://lain.bgm.tv/pic/cover/c/fc/73/306147_o4qLV.jpg
#> 34: http://lain.bgm.tv/pic/cover/c/27/3c/309002_BQIXI.jpg
#> 35: http://lain.bgm.tv/pic/cover/c/1d/86/314488_tnvn3.jpg
#> 36: http://lain.bgm.tv/pic/cover/c/b3/c9/316104_1yKUQ.jpg
#> 37: http://lain.bgm.tv/pic/cover/c/7b/ef/316795_YxeeC.jpg
#> 38: http://lain.bgm.tv/pic/cover/c/0c/80/322926_cwAAw.jpg
#> 39:    http://lain.bgm.tv/pic/cover/c/01/88/899_Q3F3X.jpg
#> 40: http://lain.bgm.tv/pic/cover/c/ae/11/278815_veMhu.jpg
#> 41: http://lain.bgm.tv/pic/cover/c/97/02/296875_3Z9k9.jpg
#> 42: http://lain.bgm.tv/pic/cover/c/bf/76/304065_4Crea.jpg
#> 43: http://lain.bgm.tv/pic/cover/c/a9/72/314447_5CQ03.jpg
#> 44: http://lain.bgm.tv/pic/cover/c/23/a2/315069_QlT2S.jpg
#> 45: http://lain.bgm.tv/pic/cover/c/8b/0f/315456_Q6m1s.jpg
#> 46:    http://lain.bgm.tv/pic/cover/c/92/97/975_YKuWd.jpg
#> 47: http://lain.bgm.tv/pic/cover/c/18/80/269565_5W5I1.jpg
#> 48: http://lain.bgm.tv/pic/cover/c/8b/00/277554_tG28z.jpg
#> 49: http://lain.bgm.tv/pic/cover/c/ac/b7/282684_0B301.jpg
#> 50: http://lain.bgm.tv/pic/cover/c/3b/af/295766_dJ65o.jpg
#> 51: http://lain.bgm.tv/pic/cover/c/14/f6/296151_b8ZGB.jpg
#> 52: http://lain.bgm.tv/pic/cover/c/6d/7c/296941_j4dd2.jpg
#> 53: http://lain.bgm.tv/pic/cover/c/86/05/315804_GPogW.jpg
#> 54: http://lain.bgm.tv/pic/cover/c/ff/d0/319894_9GIo6.jpg
#> 55: http://lain.bgm.tv/pic/cover/c/29/51/321092_yFId5.jpg
#> 56: http://lain.bgm.tv/pic/cover/c/2f/e7/321289_1M1GJ.jpg
#>                                             images.common
#>                                             images.medium
#>  1: http://lain.bgm.tv/pic/cover/m/98/e5/293193_d88nd.jpg
#>  2: http://lain.bgm.tv/pic/cover/m/42/fd/297168_hx33I.jpg
#>  3: http://lain.bgm.tv/pic/cover/m/e4/dc/297344_qhoyE.jpg
#>  4: http://lain.bgm.tv/pic/cover/m/e6/f6/301601_9TO1X.jpg
#>  5: http://lain.bgm.tv/pic/cover/m/1c/6b/315574_tLgrT.jpg
#>  6: http://lain.bgm.tv/pic/cover/m/1a/8b/320824_hQG00.jpg
#>  7: http://lain.bgm.tv/pic/cover/m/91/4d/195723_z322N.jpg
#>  8: http://lain.bgm.tv/pic/cover/m/e6/78/278031_202kC.jpg
#>  9: http://lain.bgm.tv/pic/cover/m/09/50/308961_2xti9.jpg
#> 10: http://lain.bgm.tv/pic/cover/m/56/d9/310950_5mmvt.jpg
#> 11: http://lain.bgm.tv/pic/cover/m/dd/60/316607_Qukn6.jpg
#> 12: http://lain.bgm.tv/pic/cover/m/2d/0d/278998_WRj0d.jpg
#> 13: http://lain.bgm.tv/pic/cover/m/c0/76/280615_zfF1K.jpg
#> 14: http://lain.bgm.tv/pic/cover/m/c7/3f/295017_EXE8i.jpg
#> 15: http://lain.bgm.tv/pic/cover/m/f4/44/297224_eE7lv.jpg
#> 16: http://lain.bgm.tv/pic/cover/m/d1/43/297370_Eue45.jpg
#> 17: http://lain.bgm.tv/pic/cover/m/16/c6/298698_T13k1.jpg
#> 18: http://lain.bgm.tv/pic/cover/m/fe/e5/302639_AejA9.jpg
#> 19: http://lain.bgm.tv/pic/cover/m/02/23/303172_L1cyr.jpg
#> 20: http://lain.bgm.tv/pic/cover/m/5d/db/312700_vpftp.jpg
#> 21: http://lain.bgm.tv/pic/cover/m/aa/64/316247_a6jUR.jpg
#> 22: http://lain.bgm.tv/pic/cover/m/0f/50/262897_Lfm1T.jpg
#> 23: http://lain.bgm.tv/pic/cover/m/64/81/279202_tMc9C.jpg
#> 24: http://lain.bgm.tv/pic/cover/m/6c/00/282000_Hax2m.jpg
#> 25: http://lain.bgm.tv/pic/cover/m/65/f5/293945_89GwC.jpg
#> 26: http://lain.bgm.tv/pic/cover/m/50/b5/296261_i3iaJ.jpg
#> 27: http://lain.bgm.tv/pic/cover/m/51/8d/302802_Df3oo.jpg
#> 28: http://lain.bgm.tv/pic/cover/m/b6/39/304827_Bbn7q.jpg
#> 29: http://lain.bgm.tv/pic/cover/m/83/73/252782_6PjZ1.jpg
#> 30: http://lain.bgm.tv/pic/cover/m/c8/42/292527_YtaWi.jpg
#> 31: http://lain.bgm.tv/pic/cover/m/94/8c/297396_xJK3Q.jpg
#> 32: http://lain.bgm.tv/pic/cover/m/f2/c1/302797_SZfJM.jpg
#> 33: http://lain.bgm.tv/pic/cover/m/fc/73/306147_o4qLV.jpg
#> 34: http://lain.bgm.tv/pic/cover/m/27/3c/309002_BQIXI.jpg
#> 35: http://lain.bgm.tv/pic/cover/m/1d/86/314488_tnvn3.jpg
#> 36: http://lain.bgm.tv/pic/cover/m/b3/c9/316104_1yKUQ.jpg
#> 37: http://lain.bgm.tv/pic/cover/m/7b/ef/316795_YxeeC.jpg
#> 38: http://lain.bgm.tv/pic/cover/m/0c/80/322926_cwAAw.jpg
#> 39:    http://lain.bgm.tv/pic/cover/m/01/88/899_Q3F3X.jpg
#> 40: http://lain.bgm.tv/pic/cover/m/ae/11/278815_veMhu.jpg
#> 41: http://lain.bgm.tv/pic/cover/m/97/02/296875_3Z9k9.jpg
#> 42: http://lain.bgm.tv/pic/cover/m/bf/76/304065_4Crea.jpg
#> 43: http://lain.bgm.tv/pic/cover/m/a9/72/314447_5CQ03.jpg
#> 44: http://lain.bgm.tv/pic/cover/m/23/a2/315069_QlT2S.jpg
#> 45: http://lain.bgm.tv/pic/cover/m/8b/0f/315456_Q6m1s.jpg
#> 46:    http://lain.bgm.tv/pic/cover/m/92/97/975_YKuWd.jpg
#> 47: http://lain.bgm.tv/pic/cover/m/18/80/269565_5W5I1.jpg
#> 48: http://lain.bgm.tv/pic/cover/m/8b/00/277554_tG28z.jpg
#> 49: http://lain.bgm.tv/pic/cover/m/ac/b7/282684_0B301.jpg
#> 50: http://lain.bgm.tv/pic/cover/m/3b/af/295766_dJ65o.jpg
#> 51: http://lain.bgm.tv/pic/cover/m/14/f6/296151_b8ZGB.jpg
#> 52: http://lain.bgm.tv/pic/cover/m/6d/7c/296941_j4dd2.jpg
#> 53: http://lain.bgm.tv/pic/cover/m/86/05/315804_GPogW.jpg
#> 54: http://lain.bgm.tv/pic/cover/m/ff/d0/319894_9GIo6.jpg
#> 55: http://lain.bgm.tv/pic/cover/m/29/51/321092_yFId5.jpg
#> 56: http://lain.bgm.tv/pic/cover/m/2f/e7/321289_1M1GJ.jpg
#>                                             images.medium
#>                                              images.small
#>  1: http://lain.bgm.tv/pic/cover/s/98/e5/293193_d88nd.jpg
#>  2: http://lain.bgm.tv/pic/cover/s/42/fd/297168_hx33I.jpg
#>  3: http://lain.bgm.tv/pic/cover/s/e4/dc/297344_qhoyE.jpg
#>  4: http://lain.bgm.tv/pic/cover/s/e6/f6/301601_9TO1X.jpg
#>  5: http://lain.bgm.tv/pic/cover/s/1c/6b/315574_tLgrT.jpg
#>  6: http://lain.bgm.tv/pic/cover/s/1a/8b/320824_hQG00.jpg
#>  7: http://lain.bgm.tv/pic/cover/s/91/4d/195723_z322N.jpg
#>  8: http://lain.bgm.tv/pic/cover/s/e6/78/278031_202kC.jpg
#>  9: http://lain.bgm.tv/pic/cover/s/09/50/308961_2xti9.jpg
#> 10: http://lain.bgm.tv/pic/cover/s/56/d9/310950_5mmvt.jpg
#> 11: http://lain.bgm.tv/pic/cover/s/dd/60/316607_Qukn6.jpg
#> 12: http://lain.bgm.tv/pic/cover/s/2d/0d/278998_WRj0d.jpg
#> 13: http://lain.bgm.tv/pic/cover/s/c0/76/280615_zfF1K.jpg
#> 14: http://lain.bgm.tv/pic/cover/s/c7/3f/295017_EXE8i.jpg
#> 15: http://lain.bgm.tv/pic/cover/s/f4/44/297224_eE7lv.jpg
#> 16: http://lain.bgm.tv/pic/cover/s/d1/43/297370_Eue45.jpg
#> 17: http://lain.bgm.tv/pic/cover/s/16/c6/298698_T13k1.jpg
#> 18: http://lain.bgm.tv/pic/cover/s/fe/e5/302639_AejA9.jpg
#> 19: http://lain.bgm.tv/pic/cover/s/02/23/303172_L1cyr.jpg
#> 20: http://lain.bgm.tv/pic/cover/s/5d/db/312700_vpftp.jpg
#> 21: http://lain.bgm.tv/pic/cover/s/aa/64/316247_a6jUR.jpg
#> 22: http://lain.bgm.tv/pic/cover/s/0f/50/262897_Lfm1T.jpg
#> 23: http://lain.bgm.tv/pic/cover/s/64/81/279202_tMc9C.jpg
#> 24: http://lain.bgm.tv/pic/cover/s/6c/00/282000_Hax2m.jpg
#> 25: http://lain.bgm.tv/pic/cover/s/65/f5/293945_89GwC.jpg
#> 26: http://lain.bgm.tv/pic/cover/s/50/b5/296261_i3iaJ.jpg
#> 27: http://lain.bgm.tv/pic/cover/s/51/8d/302802_Df3oo.jpg
#> 28: http://lain.bgm.tv/pic/cover/s/b6/39/304827_Bbn7q.jpg
#> 29: http://lain.bgm.tv/pic/cover/s/83/73/252782_6PjZ1.jpg
#> 30: http://lain.bgm.tv/pic/cover/s/c8/42/292527_YtaWi.jpg
#> 31: http://lain.bgm.tv/pic/cover/s/94/8c/297396_xJK3Q.jpg
#> 32: http://lain.bgm.tv/pic/cover/s/f2/c1/302797_SZfJM.jpg
#> 33: http://lain.bgm.tv/pic/cover/s/fc/73/306147_o4qLV.jpg
#> 34: http://lain.bgm.tv/pic/cover/s/27/3c/309002_BQIXI.jpg
#> 35: http://lain.bgm.tv/pic/cover/s/1d/86/314488_tnvn3.jpg
#> 36: http://lain.bgm.tv/pic/cover/s/b3/c9/316104_1yKUQ.jpg
#> 37: http://lain.bgm.tv/pic/cover/s/7b/ef/316795_YxeeC.jpg
#> 38: http://lain.bgm.tv/pic/cover/s/0c/80/322926_cwAAw.jpg
#> 39:    http://lain.bgm.tv/pic/cover/s/01/88/899_Q3F3X.jpg
#> 40: http://lain.bgm.tv/pic/cover/s/ae/11/278815_veMhu.jpg
#> 41: http://lain.bgm.tv/pic/cover/s/97/02/296875_3Z9k9.jpg
#> 42: http://lain.bgm.tv/pic/cover/s/bf/76/304065_4Crea.jpg
#> 43: http://lain.bgm.tv/pic/cover/s/a9/72/314447_5CQ03.jpg
#> 44: http://lain.bgm.tv/pic/cover/s/23/a2/315069_QlT2S.jpg
#> 45: http://lain.bgm.tv/pic/cover/s/8b/0f/315456_Q6m1s.jpg
#> 46:    http://lain.bgm.tv/pic/cover/s/92/97/975_YKuWd.jpg
#> 47: http://lain.bgm.tv/pic/cover/s/18/80/269565_5W5I1.jpg
#> 48: http://lain.bgm.tv/pic/cover/s/8b/00/277554_tG28z.jpg
#> 49: http://lain.bgm.tv/pic/cover/s/ac/b7/282684_0B301.jpg
#> 50: http://lain.bgm.tv/pic/cover/s/3b/af/295766_dJ65o.jpg
#> 51: http://lain.bgm.tv/pic/cover/s/14/f6/296151_b8ZGB.jpg
#> 52: http://lain.bgm.tv/pic/cover/s/6d/7c/296941_j4dd2.jpg
#> 53: http://lain.bgm.tv/pic/cover/s/86/05/315804_GPogW.jpg
#> 54: http://lain.bgm.tv/pic/cover/s/ff/d0/319894_9GIo6.jpg
#> 55: http://lain.bgm.tv/pic/cover/s/29/51/321092_yFId5.jpg
#> 56: http://lain.bgm.tv/pic/cover/s/2f/e7/321289_1M1GJ.jpg
#>                                              images.small
#>                                               images.grid collection.doing
#>  1: http://lain.bgm.tv/pic/cover/g/98/e5/293193_d88nd.jpg              412
#>  2: http://lain.bgm.tv/pic/cover/g/42/fd/297168_hx33I.jpg              194
#>  3: http://lain.bgm.tv/pic/cover/g/e4/dc/297344_qhoyE.jpg               20
#>  4: http://lain.bgm.tv/pic/cover/g/e6/f6/301601_9TO1X.jpg              592
#>  5: http://lain.bgm.tv/pic/cover/g/1c/6b/315574_tLgrT.jpg              390
#>  6: http://lain.bgm.tv/pic/cover/g/1a/8b/320824_hQG00.jpg               31
#>  7: http://lain.bgm.tv/pic/cover/g/91/4d/195723_z322N.jpg              270
#>  8: http://lain.bgm.tv/pic/cover/g/e6/78/278031_202kC.jpg              559
#>  9: http://lain.bgm.tv/pic/cover/g/09/50/308961_2xti9.jpg               41
#> 10: http://lain.bgm.tv/pic/cover/g/56/d9/310950_5mmvt.jpg               80
#> 11: http://lain.bgm.tv/pic/cover/g/dd/60/316607_Qukn6.jpg              202
#> 12: http://lain.bgm.tv/pic/cover/g/2d/0d/278998_WRj0d.jpg              101
#> 13: http://lain.bgm.tv/pic/cover/g/c0/76/280615_zfF1K.jpg               27
#> 14: http://lain.bgm.tv/pic/cover/g/c7/3f/295017_EXE8i.jpg              147
#> 15: http://lain.bgm.tv/pic/cover/g/f4/44/297224_eE7lv.jpg              775
#> 16: http://lain.bgm.tv/pic/cover/g/d1/43/297370_Eue45.jpg              109
#> 17: http://lain.bgm.tv/pic/cover/g/16/c6/298698_T13k1.jpg              258
#> 18: http://lain.bgm.tv/pic/cover/g/fe/e5/302639_AejA9.jpg               69
#> 19: http://lain.bgm.tv/pic/cover/g/02/23/303172_L1cyr.jpg               21
#> 20: http://lain.bgm.tv/pic/cover/g/5d/db/312700_vpftp.jpg               23
#> 21: http://lain.bgm.tv/pic/cover/g/aa/64/316247_a6jUR.jpg              986
#> 22: http://lain.bgm.tv/pic/cover/g/0f/50/262897_Lfm1T.jpg             1259
#> 23: http://lain.bgm.tv/pic/cover/g/64/81/279202_tMc9C.jpg              905
#> 24: http://lain.bgm.tv/pic/cover/g/6c/00/282000_Hax2m.jpg              538
#> 25: http://lain.bgm.tv/pic/cover/g/65/f5/293945_89GwC.jpg              138
#> 26: http://lain.bgm.tv/pic/cover/g/50/b5/296261_i3iaJ.jpg              270
#> 27: http://lain.bgm.tv/pic/cover/g/51/8d/302802_Df3oo.jpg              102
#> 28: http://lain.bgm.tv/pic/cover/g/b6/39/304827_Bbn7q.jpg              343
#> 29: http://lain.bgm.tv/pic/cover/g/83/73/252782_6PjZ1.jpg              551
#> 30: http://lain.bgm.tv/pic/cover/g/c8/42/292527_YtaWi.jpg              591
#> 31: http://lain.bgm.tv/pic/cover/g/94/8c/297396_xJK3Q.jpg              271
#> 32: http://lain.bgm.tv/pic/cover/g/f2/c1/302797_SZfJM.jpg               65
#> 33: http://lain.bgm.tv/pic/cover/g/fc/73/306147_o4qLV.jpg              266
#> 34: http://lain.bgm.tv/pic/cover/g/27/3c/309002_BQIXI.jpg               19
#> 35: http://lain.bgm.tv/pic/cover/g/1d/86/314488_tnvn3.jpg               53
#> 36: http://lain.bgm.tv/pic/cover/g/b3/c9/316104_1yKUQ.jpg               12
#> 37: http://lain.bgm.tv/pic/cover/g/7b/ef/316795_YxeeC.jpg               27
#> 38: http://lain.bgm.tv/pic/cover/g/0c/80/322926_cwAAw.jpg               93
#> 39:    http://lain.bgm.tv/pic/cover/g/01/88/899_Q3F3X.jpg             3601
#> 40: http://lain.bgm.tv/pic/cover/g/ae/11/278815_veMhu.jpg              581
#> 41: http://lain.bgm.tv/pic/cover/g/97/02/296875_3Z9k9.jpg               33
#> 42: http://lain.bgm.tv/pic/cover/g/bf/76/304065_4Crea.jpg              594
#> 43: http://lain.bgm.tv/pic/cover/g/a9/72/314447_5CQ03.jpg               66
#> 44: http://lain.bgm.tv/pic/cover/g/23/a2/315069_QlT2S.jpg              680
#> 45: http://lain.bgm.tv/pic/cover/g/8b/0f/315456_Q6m1s.jpg              223
#> 46:    http://lain.bgm.tv/pic/cover/g/92/97/975_YKuWd.jpg             6031
#> 47: http://lain.bgm.tv/pic/cover/g/18/80/269565_5W5I1.jpg               59
#> 48: http://lain.bgm.tv/pic/cover/g/8b/00/277554_tG28z.jpg              545
#> 49: http://lain.bgm.tv/pic/cover/g/ac/b7/282684_0B301.jpg              417
#> 50: http://lain.bgm.tv/pic/cover/g/3b/af/295766_dJ65o.jpg              111
#> 51: http://lain.bgm.tv/pic/cover/g/14/f6/296151_b8ZGB.jpg              104
#> 52: http://lain.bgm.tv/pic/cover/g/6d/7c/296941_j4dd2.jpg              149
#> 53: http://lain.bgm.tv/pic/cover/g/86/05/315804_GPogW.jpg               23
#> 54: http://lain.bgm.tv/pic/cover/g/ff/d0/319894_9GIo6.jpg               61
#> 55: http://lain.bgm.tv/pic/cover/g/29/51/321092_yFId5.jpg               20
#> 56: http://lain.bgm.tv/pic/cover/g/2f/e7/321289_1M1GJ.jpg               22
#>                                               images.grid collection.doing