Get summary of user's collection

Get summary of user's collection

bgm_latest(
  username,
  type = c("anime", "book", "game", "music", "real"),
  max = 25,
  format = c("list", "table", "default"),
  key = "bgm2725b1fa2fea5582"
)

Arguments

username
User name a user id.

type
Category type, anime, book, game, music and real are available.

max
Max result number, default is the 25 (Maximum value).

format
Result will be a list of json data when set to list (default), response() object when set to default, and list of data.table when set to table.

key
Your application key (ID). bangumir already comes with a key/secret but you are welcome to swap out with our own. Since these keys are shipped with the package, there is a small chance they could be voided if someone abuses the key. If you plan to use this in production, or for an internal tool, the recommended practice is to create a new application on Bangumi and use those keys for your purposes.

Value

A list of json data in list, response() object, or a data.table contains the latest collections.

Examples

result <- bgm_latest(username = "sai", format = "table")

# Result of watching
result$do

#>     subject_id                  subject.url subject.type
#>  1:     217573 http://bgm.tv/subject/217573        anime
#>  2:     273843 http://bgm.tv/subject/273843        anime
#>  3:     271724 http://bgm.tv/subject/271724        anime
#>  4:     274613 http://bgm.tv/subject/274613        anime
#>  5:     207195 http://bgm.tv/subject/207195        anime
#>  6:     183878 http://bgm.tv/subject/183878        anime
#>  7:      45336  http://bgm.tv/subject/45336        anime
#>  8:     174142 http://bgm.tv/subject/174142        anime
#>  9:        324    http://bgm.tv/subject/324        anime
#> 10:        893    http://bgm.tv/subject/893        anime
#> 11:          8      http://bgm.tv/subject/8        anime
#> 12:      85316  http://bgm.tv/subject/85316        anime
#> 13:     140001 http://bgm.tv/subject/140001        anime
#> 14:     144087 http://bgm.tv/subject/144087        anime
#> 15:        843    http://bgm.tv/subject/843        anime
#> 16:     139408 http://bgm.tv/subject/139408        anime
#> 17:      72941  http://bgm.tv/subject/72941        anime
#> 18:     104521 http://bgm.tv/subject/104521        anime
#> 19:       1453   http://bgm.tv/subject/1453        anime
#> 20:      93739  http://bgm.tv/subject/93739        anime
#> 21:      79973  http://bgm.tv/subject/79973        anime
#> 22:      51928  http://bgm.tv/subject/51928        anime
#> 23:      55113  http://bgm.tv/subject/55113        anime
#> 24:      50331  http://bgm.tv/subject/50331        anime
#> 25:       1266   http://bgm.tv/subject/1266        anime
#>     subject_id                  subject.url subject.type
#>                         subject.name                subject.name_cn
#>  1:           リラックマとカオルさん                     轻松熊与薰
#>  2:                         BEASTARS                       BEASTARS
#>  3:           ダンベル何キロ持てる？ 肌肉少女：哑铃，能举多少公斤？
#>  4:         Love, Death &amp; Robots          爱、死亡 &amp; 机器人
#>  5:                      ゆるキャン△                      摇曳露营△
#>  6: ヴァイオレット・エヴァーガーデン                 紫罗兰永恒花园
#>  7:           Gravity Falls Season 1                怪诞小镇 第一季
#>  8:                         舟を編む                         编舟记
#>  9:   攻殻機動隊 STAND ALONE COMPLEX 攻壳机动队 STAND ALONE COMPLEX
#> 10:                         灰羽連盟                       灰羽联盟
#> 11:  コードギアス 反逆のルルーシュR2      Code Geass 反叛的鲁路修R2
#> 12:          Adventure Time season 5                探险活宝 第五季
#> 13:      Re:ゼロから始める異世界生活       Re：从零开始的异世界生活
#> 14:         Bojack Horseman Season 2                    马男波杰克2
#> 15:                       妄想代理人                     妄想代理人
#> 16:                       おそ松さん                           阿松
#> 17:                       キルラキル                       斩服少女
#> 18:          Adventure Time season 6                探险活宝 第六季
#> 19:                   少女革命ウテナ                       少女革命
#> 20:           ピンポン THE ANIMATION                           乒乓
#> 21:            Mickey Mouse Season 1              米老鼠2013 第一季
#> 22:            とある科学の超電磁砲S              某科学的超电磁炮S
#> 23:                 たまこまーけっと                       玉子市场
#> 24:       中二病でも恋がしたい！Lite         中二病也要谈恋爱！Lite
#> 25:           交響詩篇エウレカセブン                       交响诗篇
#>                         subject.name                subject.name_cn
#>                                      subject.images.large
#>  1: http://lain.bgm.tv/pic/cover/l/c0/de/217573_trel5.jpg
#>  2: http://lain.bgm.tv/pic/cover/l/f6/f2/273843_YMdDZ.jpg
#>  3: http://lain.bgm.tv/pic/cover/l/b7/b7/271724_32Jou.jpg
#>  4: http://lain.bgm.tv/pic/cover/l/68/b8/274613_sx1iX.jpg
#>  5: http://lain.bgm.tv/pic/cover/l/18/bc/207195_OOyqz.jpg
#>  6: http://lain.bgm.tv/pic/cover/l/1e/e2/183878_Fef1o.jpg
#>  7:  http://lain.bgm.tv/pic/cover/l/a4/34/45336_ftJ2J.jpg
#>  8: http://lain.bgm.tv/pic/cover/l/71/ff/174142_HirNZ.jpg
#>  9:    http://lain.bgm.tv/pic/cover/l/f2/fc/324_psuXk.jpg
#> 10:    http://lain.bgm.tv/pic/cover/l/d5/6b/893_gEo72.jpg
#> 11:      http://lain.bgm.tv/pic/cover/l/c9/f0/8_wK0z3.jpg
#> 12:  http://lain.bgm.tv/pic/cover/l/69/06/85316_aW8c9.jpg
#> 13: http://lain.bgm.tv/pic/cover/l/cb/78/140001_Ew1mo.jpg
#> 14: http://lain.bgm.tv/pic/cover/l/4c/8f/144087_E3I19.jpg
#> 15:    http://lain.bgm.tv/pic/cover/l/3d/8e/843_6wj9L.jpg
#> 16: http://lain.bgm.tv/pic/cover/l/c0/fe/139408_yqY6B.jpg
#> 17:  http://lain.bgm.tv/pic/cover/l/d5/d7/72941_q1DiW.jpg
#> 18: http://lain.bgm.tv/pic/cover/l/00/b9/104521_6m21b.jpg
#> 19:   http://lain.bgm.tv/pic/cover/l/53/6a/1453_iZIOZ.jpg
#> 20:  http://lain.bgm.tv/pic/cover/l/1e/63/93739_TZ9dS.jpg
#> 21:  http://lain.bgm.tv/pic/cover/l/0a/d5/79973_hRTf0.jpg
#> 22:  http://lain.bgm.tv/pic/cover/l/99/8c/51928_wuG5D.jpg
#> 23:  http://lain.bgm.tv/pic/cover/l/f8/7f/55113_TR5Is.jpg
#> 24:  http://lain.bgm.tv/pic/cover/l/8f/2c/50331_KWKjN.jpg
#> 25:   http://lain.bgm.tv/pic/cover/l/8e/2c/1266_h6m69.jpg
#>                                      subject.images.large