Get subject information.

Get subject information.

bgm_subject(
  id = NULL,
  responseGroup = c("small", "medium", "large"),
  format = c("list", "default")
)

Arguments

id
Subject id.

responseGroup
Response of details type is set to be small (simplified mode), medium and large are also available.

format
Result will be a list of json data when set to list (default), response() object when set to default.

Value

A list of response content or a response() object according to the format option.

Examples

bgm_subject(10380)

#> $id
#> [1] 10380
#> 
#> $url
#> [1] "http://bgm.tv/subject/10380"
#> 
#> $type
#> [1] 2
#> 
#> $name
#> [1] "STEINS;GATE"
#> 
#> $name_cn
#> [1] "命运石之门"
#> 
#> $summary
#> [1] "       故事发生在「CHAOS;HEAD」的“涩谷崩坏”事件一年半之后的世界，而舞台则从涩谷转移到了秋叶原。主角冈部伦太郎是一位深度中二病的大学生，时常幻想自己身肩重任，并自称“狂气的疯狂科学家・凤凰院凶真”，不过说到底其作为不过就是在名为“未来道具研究所”中与两个伙伴开发着奇奇怪怪又不切实际的东西。然而，这样的他们却在偶然间发明出可以把电子讯息传送过去的时间机器。在他们对未来及过去知道得越多的同时，却不知道危难正渐渐临到自己身上……"
#> 
#> $eps
#> [1] 25
#> 
#> $eps_count
#> [1] 25
#> 
#> $air_date
#> [1] "2011-04-06"
#> 
#> $air_weekday
#> [1] 3
#> 
#> $rating
#> $rating$total
#> [1] 14480
#> 
#> $rating$count
#> $rating$count$`10`
#> [1] 3037
#> 
#> $rating$count$`9`
#> [1] 5447
#> 
#> $rating$count$`8`
#> [1] 4002
#> 
#> $rating$count$`7`
#> [1] 1400
#> 
#> $rating$count$`6`
#> [1] 375
#> 
#> $rating$count$`5`
#> [1] 125
#> 
#> $rating$count$`4`
#> [1] 31
#> 
#> $rating$count$`3`
#> [1] 8
#> 
#> $rating$count$`2`
#> [1] 10
#> 
#> $rating$count$`1`
#> [1] 45
#> 
#> 
#> $rating$score
#> [1] 8.6
#> 
#> 
#> $rank
#> [1] 27
#> 
#> $images
#> $images$large
#> [1] "http://lain.bgm.tv/pic/cover/l/a9/79/10380_YwP4R.jpg"
#> 
#> $images$common
#> [1] "http://lain.bgm.tv/pic/cover/c/a9/79/10380_YwP4R.jpg"
#> 
#> $images$medium
#> [1] "http://lain.bgm.tv/pic/cover/m/a9/79/10380_YwP4R.jpg"
#> 
#> $images$small
#> [1] "http://lain.bgm.tv/pic/cover/s/a9/79/10380_YwP4R.jpg"
#> 
#> $images$grid
#> [1] "http://lain.bgm.tv/pic/cover/g/a9/79/10380_YwP4R.jpg"
#> 
#> 
#> $collection
#> $collection$wish
#> [1] 1902
#> 
#> $collection$collect
#> [1] 18065
#> 
#> $collection$doing
#> [1] 1259
#> 
#> $collection$on_hold
#> [1] 678
#> 
#> $collection$dropped
#> [1] 256
#> 
#>