Get subject episode information.

Get subject episode information.

bgm_subject_ep(id = NULL, format = c("list", "table", "default"))

Arguments

id
Subject id.

format
Result will be a list of json data when set to list (default), response() object when set to default. Set to table will convert the episode list to a data.frame for ease of use.

Value

A list of response content or a response() object according to the format option.

Examples

bgm_subject_ep(10380, "table")

#> $id
#> [1] 10380
#> 
#> $url
#> [1] "http://bgm.tv/subject/10380"
#> 
#> $type
#> [1] 2
#> 
#> $name
#> [1] "STEINS;GATE"
#> 
#> $name_cn
#> [1] "命运石之门"
#> 
#> $summary
#> [1] ""
#> 
#> $air_date
#> [1] ""
#> 
#> $air_weekday
#> [1] 0
#> 
#> $images
#> $images$large
#> [1] "http://lain.bgm.tv/pic/cover/l/a9/79/10380_YwP4R.jpg"
#> 
#> $images$common
#> [1] "http://lain.bgm.tv/pic/cover/c/a9/79/10380_YwP4R.jpg"
#> 
#> $images$medium
#> [1] "http://lain.bgm.tv/pic/cover/m/a9/79/10380_YwP4R.jpg"
#> 
#> $images$small
#> [1] "http://lain.bgm.tv/pic/cover/s/a9/79/10380_YwP4R.jpg"
#> 
#> $images$grid
#> [1] "http://lain.bgm.tv/pic/cover/g/a9/79/10380_YwP4R.jpg"
#> 
#> 
#> $eps
#>         id                     url type sort
#>  1:  84517  http://bgm.tv/ep/84517    0    1
#>  2:  84518  http://bgm.tv/ep/84518    0    2
#>  3:  84519  http://bgm.tv/ep/84519    0    3
#>  4:  84520  http://bgm.tv/ep/84520    0    4
#>  5:  84521  http://bgm.tv/ep/84521    0    5
#>  6:  84522  http://bgm.tv/ep/84522    0    6
#>  7:  84523  http://bgm.tv/ep/84523    0    7
#>  8:  84524  http://bgm.tv/ep/84524    0    8
#>  9:  84525  http://bgm.tv/ep/84525    0    9
#> 10:  84526  http://bgm.tv/ep/84526    0   10
#> 11:  84527  http://bgm.tv/ep/84527    0   11
#> 12:  84715  http://bgm.tv/ep/84715    0   12
#> 13:  91560  http://bgm.tv/ep/91560    0   13
#> 14:  91561  http://bgm.tv/ep/91561    0   14
#> 15:  91562  http://bgm.tv/ep/91562    0   15
#> 16:  91563  http://bgm.tv/ep/91563    0   16
#> 17:  91564  http://bgm.tv/ep/91564    0   17
#> 18:  91565  http://bgm.tv/ep/91565    0   18
#> 19:  91566  http://bgm.tv/ep/91566    0   19
#> 20:  91567  http://bgm.tv/ep/91567    0   20
#> 21:  91568  http://bgm.tv/ep/91568    0   21
#> 22:  91569  http://bgm.tv/ep/91569    0   22
#> 23:  91570  http://bgm.tv/ep/91570    0   23
#> 24: 589408 http://bgm.tv/ep/589408    1   23
#> 25:  91571  http://bgm.tv/ep/91571    0   24
#> 26: 150204 http://bgm.tv/ep/150204    1   25
#>         id                     url type sort
#>                                                      name
#>  1:             始まりと終わりのプロローグ-Turning Point-
#>  2:            時間跳躍のパラノイア-Time Travel Paranoia-
#>  3:         並列過程のパラノイア-Parallel World Paranoia-
#>  4:          空理彷徨のランデヴー-Interpreter Rendezvous-
#>  5:             電荷衝突のランデヴー-Starmine Rendezvous-
#>  6: 蝶翼のダイバージェンス-Butterfly Effect's Divergence-
#>  7:        断層のダイバージェンス-Divergence Singularity-
#>  8:         夢幻のホメオスタシス-Chaos Theory Homeostasis
#>  9:        幻相のホメオスタシス-Chaos Theory Homeostasis-
#> 10:        相生のホメオスタシス-Chaos Theory Homeostasis-
#> 11:              時空境界のドグマ-Dogma in Event Horizon-
#> 12:                 静止限界のドグマ-Dogma in Ergosphere-
#> 13:            形而上のネクローシス-Metaphysics Necrosis-
#> 14:             形而下のネクローシス-Physically Necrosis-
#> 15:           亡環上のネクローシス-Missing Link Necrosis-
#> 16:            不可逆のネクローシス-Sacrificial Necrosis-
#> 17:        虚像歪曲のコンプレックス-Sacrificial Necrosis-
#> 18:       自己相似のアンドロギュヌス-Fractal Androgynous-
#> 19:             無限連鎖のアポトーシス-Endless Apoptosis-
#> 20:           怨嗟断絶のアポトーシス-Finalize Apoptosis -
#> 21:                      因果律のメルト-Paradox Meltdown-
#> 22:                      存在了解のメルト-Being Meltdown-
#> 23:    境界面上のシュタインズゲート-Open The Steins Gate-
#> 24:        (β) 境界面上のミッシングリンク-Divide by Zero-
#> 25:      終わりと始まりのプロローグ-Open The Steins Gate-
#> 26:           横行跋扈のポリオマニア-Egoistic Poriomania-
#>                                                      name
#>                                        name_cn duration    airdate comment
#>  1:                始与终的序章-Turning Point- 00:23:40 2011-04-05      91
#>  2:     时间跳跃的偏执狂-Time Travel Paranoia- 00:23:40 2011-04-12      37
#>  3:  并列过程的偏执狂-Parallel World Paranoia- 00:23:40 2011-04-19      22
#>  4:     空谈彷徨的交汇-Interpreter Rendezvous- 00:23:40 2011-04-26      27
#>  5:        电荷冲突的交汇-Starmine Rendezvous- 00:23:40 2011-05-03      39
#>  6:  蝶翼的发散-Butterfly Effect's Divergence- 00:23:40 2011-05-10      27
#>  7:         断层的发散-Divergence Singularity- 00:23:40 2011-05-17      20
#>  8:    梦幻的恒定状态-Chaos Theory Homeostasis 00:23:40 2011-05-24      43
#>  9:   幻相的恒定状态-Chaos Theory Homeostasis- 00:23:40 2011-05-31      55
#> 10:   相生的恒定状态-Chaos Theory Homeostasis- 00:23:40 2011-06-07      34
#> 11:     时空境界的教义-Dogma in Event Horizon- 00:23:40 2011-06-14      37
#> 12:        静止界限的教义-Dogma in Ergosphere- 00:23:40 2011-06-21      57
#> 13:         形而上的坏死-Metaphysics Necrosis- 00:23:40 2011-06-28      75
#> 14:          形而下的坏死-Physically Necrosis- 00:23:40 2011-07-05      61
#> 15:        亡环上的坏死-Missing Link Necrosis- 00:23:40 2011-07-12      31
#> 16:         不可逆的坏死-Sacrificial Necrosis- 00:23:40 2011-07-19      58
#> 17:     虚像歪曲的情意综-Sacrificial Necrosis- 00:23:40 2011-07-26      40
#> 18:    自己相似的两性同体-Fractal Androgynous- 00:23:40 2011-08-02      49
#> 19:      无限连锁的细胞凋亡-Endless Apoptosis- 00:23:40 2011-08-09      33
#> 20:    怨恨灭绝的细胞凋亡-Finalize Apoptosis - 00:23:40 2011-08-16      47
#> 21:             因果律的消散-Paradox Meltdown- 00:23:40 2011-08-23      27
#> 22:             存在了解的消散-Being Meltdown- 00:23:40 2011-08-30      89
#> 23: 境界面上的斯坦因之门-Open The Steins Gate- 00:23:40 2011-09-06      64
#> 24:         境界面上的缺失之环-Divide by Zero- 00:23:40 2015-12-02      18
#> 25:         终与始的序章-Open The Steins Gate- 00:23:40 2011-09-13      81
#> 26:    横行跋扈的浪荡之徒-Egoistic Poriomania- 00:24:40 2012-02-22      58
#>                                        name_cn duration    airdate comment
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             desc
#>  1:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#>  2:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#>  3:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#>  4:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#>  5:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#>  6:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#>  7:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#>  8:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#>  9:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 10:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 11:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 12:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 13:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 14:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 15:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 16:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 17:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 18:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 19:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 20:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 21:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 22:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 23:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 24: 【改変版 23話】\r\n絶望に押しつぶされた岡部を気遣うまゆり。岡部は、紅莉栖を救うことを諦め、\r\n世界の命運を背負うその重責から解放される。だが、そこは紅莉栖との出会い、\r\nラボメンとして過ごした日々が、なかったことになっている世界線であった。\r\n罪悪感を抱えたまま過ごす岡部に、"β世界線"はどのような未来を見せるのか？\r\n\r\n今回の改変により物語は第23話で完結となり、第24話は放映されません。\r\nこの"β世界線"での物語は、PlayStation 4/3/Vita『STEINS;GATE 0』へとつながっていきます。\r\nなお、再放送第24話が放送予定だった時間には、\r\n特別番組「Making of STEINS;GATE 0 ～終わりと始まりのフラグメント～」をお届けいたします。
#> 25:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> 26:                                                                                                                         2012年2月22日発売予定\r\n\r\n初回限定版Blu-rayは、原作キャラクターデザインのhuke氏による新規描き下ろしBOXに加え、キャラソン／オリジナルドラマ／サントラ／トークなど、毎巻異なるコンテンツを収録した「未来ガジェットコンパクトディスク」、本編の一部をランダムにカットした特製フィルム「時を超えた郷愁への記憶〜ノスタルジア・フィルム〜」など、豪華特典付きです！\r\n◆収録話数\r\n\r\n第23話「境界面上のシュタインズゲート」\r\n\r\n第24話「終わりと始まりのプロローグ」\r\n\r\nスペシャル「横行跋扈のポリオマニア」
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             desc
#>     status
#>  1:    Air
#>  2:    Air
#>  3:    Air
#>  4:    Air
#>  5:    Air
#>  6:    Air
#>  7:    Air
#>  8:    Air
#>  9:    Air
#> 10:    Air
#> 11:    Air
#> 12:    Air
#> 13:    Air
#> 14:    Air
#> 15:    Air
#> 16:    Air
#> 17:    Air
#> 18:    Air
#> 19:    Air
#> 20:    Air
#> 21:    Air
#> 22:    Air
#> 23:    Air
#> 24:    Air
#> 25:    Air
#> 26:    Air
#>     status
#>