Get user's collection

Get user's collection

bgm_user_collection(
  username,
  cat = c("watching", "all_watching"),
  ids = NULL,
  responseGroup = c("medium", "small"),
  format = c("list", "table", "default")
)

Arguments

username
User name a user id.

cat
Collection category, watching for anime and real, all_watching for anime, real and book.

ids
Numeric vector of subject id.

responseGroup
Response of details type is set to be medium, setting small for simplified mode.

format
Result will be a list of json data when set to list (default), response() object when set to default, and a data.table when set to table.

Examples

bgm_user_collection(username = "sai", format = "table")

#>                                      name subject_id ep_status vol_status
#>  1:                        危険なビーナス     313504         6          0
#>  2:                リラックマとカオルさん     217573         2          0
#>  3:              L'amica geniale Season 1     267251         4          0
#>  4:                     Watchmen Season 1     282499         3          0
#>  5:                              BEASTARS     273843         4          0
#>  6: The End of the F***ing World Season 1     228870         2          0
#>  7:                      Fleabag Season 1     190930         1          0
#>  8:                   Mindhunter Season 1     227600         8          0
#>  9:                ダンベル何キロ持てる？     271724         1          0
#> 10:              わたし、定時で帰ります。     272458         1          0
#> 11:              Love, Death &amp; Robots     274613        13          0
#> 12:   The Man in the High Castle Season 1     123006         2          0
#> 13:                          半分、青い。     209963        85          0
#> 14:     3年A組 ―今から皆さんは、人質です―     266647         0          0
#> 15:                    義母と娘のブルース     245130         3          0
#> 16:                  Yellowstone Season 1     251578         0          0
#> 17:                       Patrick Melrose     210409         4          0
#> 18:                        ブラックペアン     236722         1          0
#> 19:                              正義のセ     235836         0          0
#> 20:                                 anone     231037         3          0
#> 21:                           ゆるキャン△     207195         6          0
#> 22:                        監獄のお姫さま     220722         5          0
#> 23:      ヴァイオレット・エヴァーガーデン     183878         1          0
#> 24:                Gravity Falls Season 1      45336        14          0
#> 25:                              舟を編む     174142         1          0
#> 26:                        東京タラレバ娘     192356         2          0
#> 27:        攻殻機動隊 STAND ALONE COMPLEX        324         3          0
#> 28:    地味にスゴイ！校閲ガール・河野悦子     189708         9          0
#> 29:                            あさが来た     128526        74          0
#> 30:                              灰羽連盟        893         1          0
#> 31:       コードギアス 反逆のルルーシュR2          8        18          0
#> 32:               Adventure Time season 5      85316         9          0
#> 33:           Re:ゼロから始める異世界生活     140001        20          0
#> 34:              Bojack Horseman Season 2     144087         4          0
#> 35:                            妄想代理人        843         1          0
#> 36:                   Versailles Season 1     180945         2          0
#> 37:        Thunderbolt Fantasy 東離劍遊紀     168395         1          0
#> 38:            Game of Thrones (Season 1)      15877         2          0
#> 39:                            おそ松さん     139408        19          0
#> 40:                    のだめカンタービレ       1352         4          0
#> 41:                     Sense8 (Season 1)     136376        10          0
#> 42:                          アイムホーム     126014         4          0
#> 43:                          Breaking Bad       5337         4          0
#> 44:                                  民王     139201         3          0
#> 45:                             Dr.倫太郎     127132         3          0
#> 46:             House of Cards (Season 3)     119224         1          0
#> 47:                        ごめんね青春！     110248         1          0
#> 48:                            キルラキル      72941         8          0
#> 49:               Adventure Time season 6     104521         7          0
#> 50:                        少女革命ウテナ       1453         2          0
#>                                      name subject_id ep_status vol_status
#>      lasttouch subject.id                  subject.url subject.type
#>  1: 1605621718     313504 http://bgm.tv/subject/313504            6
#>  2: 1594477132     217573 http://bgm.tv/subject/217573            2
#>  3: 1592128019     267251 http://bgm.tv/subject/267251            6
#>  4: 1580313750     282499 http://bgm.tv/subject/282499            6
#>  5: 1578488802     273843 http://bgm.tv/subject/273843            2
#>  6: 1574782007     228870 http://bgm.tv/subject/228870            6
#>  7: 1569906784     190930 http://bgm.tv/subject/190930            6
#>  8: 1569649647     227600 http://bgm.tv/subject/227600            6
#>  9: 1563885005     271724 http://bgm.tv/subject/271724            2
#> 10: 1558958670     272458 http://bgm.tv/subject/272458            6
#> 11: 1556347899     274613 http://bgm.tv/subject/274613            2
#> 12: 1550902047     123006 http://bgm.tv/subject/123006            6
#> 13: 1550410646     209963 http://bgm.tv/subject/209963            6
#> 14: 1547382365     266647 http://bgm.tv/subject/266647            6
#> 15: 1538963590     245130 http://bgm.tv/subject/245130            6
#> 16: 1535870762     251578 http://bgm.tv/subject/251578            6
#> 17: 1530420192     210409 http://bgm.tv/subject/210409            6
#> 18: 1525500244     236722 http://bgm.tv/subject/236722            6
#> 19: 1524117128     235836 http://bgm.tv/subject/235836            6
#> 20: 1523104072     231037 http://bgm.tv/subject/231037            6
#> 21: 1522477122     207195 http://bgm.tv/subject/207195            2
#> 22: 1518240007     220722 http://bgm.tv/subject/220722            6
#> 23: 1515608943     183878 http://bgm.tv/subject/183878            2
#> 24: 1506664022      45336  http://bgm.tv/subject/45336            2
#> 25: 1501564974     174142 http://bgm.tv/subject/174142            2
#> 26: 1486885690     192356 http://bgm.tv/subject/192356            6
#> 27: 1482840910        324    http://bgm.tv/subject/324            2
#> 28: 1482299592     189708 http://bgm.tv/subject/189708            6
#> 29: 1482114057     128526 http://bgm.tv/subject/128526            6
#> 30: 1475173973        893    http://bgm.tv/subject/893            2
#> 31: 1473587733          8      http://bgm.tv/subject/8            2
#> 32: 1472474478      85316  http://bgm.tv/subject/85316            2
#> 33: 1472391294     140001 http://bgm.tv/subject/140001            2
#> 34: 1471535244     144087 http://bgm.tv/subject/144087            2
#> 35: 1471266640        843    http://bgm.tv/subject/843            2
#> 36: 1471159704     180945 http://bgm.tv/subject/180945            6
#> 37: 1468044163     168395 http://bgm.tv/subject/168395            6
#> 38: 1465312871      15877  http://bgm.tv/subject/15877            6
#> 39: 1463815719     139408 http://bgm.tv/subject/139408            2
#> 40: 1453544569       1352   http://bgm.tv/subject/1352            6
#> 41: 1443339227     136376 http://bgm.tv/subject/136376            6
#> 42: 1440773996     126014 http://bgm.tv/subject/126014            6
#> 43: 1440517865       5337   http://bgm.tv/subject/5337            6
#> 44: 1440435184     139201 http://bgm.tv/subject/139201            6
#> 45: 1432828713     127132 http://bgm.tv/subject/127132            6
#> 46: 1427542700     119224 http://bgm.tv/subject/119224            6
#> 47: 1419755966     110248 http://bgm.tv/subject/110248            6
#> 48: 1410169383      72941  http://bgm.tv/subject/72941            2
#> 49: 1408271732     104521 http://bgm.tv/subject/104521            2
#> 50: 1404192670       1453   http://bgm.tv/subject/1453            2
#>      lasttouch subject.id                  subject.url subject.type
#>                              subject.name                subject.name_cn
#>  1:                        危険なビーナス                   危险的维纳斯
#>  2:                リラックマとカオルさん                     轻松熊与薰
#>  3:              L'amica geniale Season 1            我的天才女友 第一季
#>  4:                     Watchmen Season 1                  守望者 第一季
#>  5:                              BEASTARS                       BEASTARS
#>  6: The End of the F***ing World Season 1            去他妈的世界 第一季
#>  7:                      Fleabag Season 1                伦敦生活 第一季
#>  8:                   Mindhunter Season 1                心灵猎人 第一季
#>  9:                ダンベル何キロ持てる？ 肌肉少女：哑铃，能举多少公斤？
#> 10:              わたし、定時で帰ります。                 我，到点下班。
#> 11:              Love, Death &amp; Robots          爱、死亡 &amp; 机器人
#> 12:   The Man in the High Castle Season 1                 高堡奇人第一季
#> 13:                          半分、青い。                     一半，蓝色
#> 14:     3年A組 ―今から皆さんは、人質です―    3年A班 -现在开始，全员人质-
#> 15:                    義母と娘のブルース               继母与女儿的蓝调
#> 16:                  Yellowstone Season 1                    黄石 第一季
#> 17:                       Patrick Melrose                       梅尔罗斯
#> 18:                        ブラックペアン                     黑色止血钳
#> 19:                              正義のセ                       正义之凛
#> 20:                                 anone                       跟你说哦
#> 21:                           ゆるキャン△                      摇曳露营△
#> 22:                        監獄のお姫さま               监狱里的公主大人
#> 23:      ヴァイオレット・エヴァーガーデン                 紫罗兰永恒花园
#> 24:                Gravity Falls Season 1                怪诞小镇 第一季
#> 25:                              舟を編む                         编舟记
#> 26:                        東京タラレバ娘                   东京白日梦女
#> 27:        攻殻機動隊 STAND ALONE COMPLEX 攻壳机动队 STAND ALONE COMPLEX
#> 28:    地味にスゴイ！校閲ガール・河野悦子  平凡不简单！校阅女孩·河野悦子
#> 29:                            あさが来た                       阿浅来了
#> 30:                              灰羽連盟                       灰羽联盟
#> 31:       コードギアス 反逆のルルーシュR2      Code Geass 反叛的鲁路修R2
#> 32:               Adventure Time season 5                探险活宝 第五季
#> 33:           Re:ゼロから始める異世界生活       Re：从零开始的异世界生活
#> 34:              Bojack Horseman Season 2                    马男波杰克2
#> 35:                            妄想代理人                     妄想代理人
#> 36:                   Versailles Season 1                  凡尔赛 第一季
#> 37:        Thunderbolt Fantasy 東離劍遊紀 Thunderbolt Fantasy 东离剑游纪
#> 38:            Game of Thrones (Season 1)              权力的游戏 第一季
#> 39:                            おそ松さん                           阿松
#> 40:                    のだめカンタービレ                     交响情人梦
#> 41:                     Sense8 (Season 1)              超感八人组 第一季
#> 42:                          アイムホーム                       I’m Home
#> 43:                          Breaking Bad                       绝命毒师
#> 44:                                  民王                           民王
#> 45:                             Dr.倫太郎                      Dr.伦太郎
#> 46:             House of Cards (Season 3)               纸牌屋（第三季）
#> 47:                        ごめんね青春！                   对不起青春！
#> 48:                            キルラキル                       斩服少女
#> 49:               Adventure Time season 6                探险活宝 第六季
#> 50:                        少女革命ウテナ                       少女革命
#>                              subject.name                subject.name_cn
#>     subject.summary subject.air_date subject.air_weekday
#>  1:                       2020-10-11                   7
#>  2:                       2019-04-19                   5
#>  3:                       2018-11-18                   7
#>  4:                       0000-00-00                   1
#>  5:                       2019-10-09                   3
#>  6:                       2017-10-24                   2
#>  7:                       2016-07-21                   4
#>  8:                       2017-10-13                   5
#>  9:                       2019-07-03                   3
#> 10:                       2019-04-16                   2
#> 11:                       2019-03-15                   5
#> 12:                       2015-11-20                   5
#> 13:                       2018-04-02                   1
#> 14:                       2019-01-06                   7
#> 15:                       2018-07-10                   2
#> 16:                       2018-06-20                   3
#> 17:                       2018-05-12                   6
#> 18:                       2018-04-22                   7
#> 19:                       2018-04-11                   3
#> 20:                       2018-01-10                   3
#> 21:                       2018-01-04                   4
#> 22:                       2017-10-17                   2
#> 23:                       2018-01-10                   3
#> 24:                       2012-06-16                   6
#> 25:                       2016-10-13                   4
#> 26:                       2017-01-18                   3
#> 27:                       2002-10-01                   2
#> 28:                       2016-10-05                   3
#> 29:                       2015-09-28                   1
#> 30:                       2002-10-09                   3
#> 31:                       2008-04-06                   7
#> 32:                       2012-11-12                   1
#> 33:                       2016-04-03                   7
#> 34:                       2015-07-17                   5
#> 35:                       2004-02-02                   6
#> 36:                       2015-11-16                   1
#> 37:                       2016-07-08                   5
#> 38:                       2011-04-17                   7
#> 39:                       2015-10-05                   1
#> 40:                       2006-10-16                   1
#> 41:                       2015-06-05                   5
#> 42:                       2015-04-16                   4
#> 43:                       2008-01-20                   7
#> 44:                       2015-07-24                   5
#> 45:                       2015-04-15                   3
#> 46:                       2015-02-27                   5
#> 47:                       2014-10-12                   7
#> 48:                       2013-10-03                   4
#> 49:                       2014-04-21                   1
#> 50:                       1997-04-02                   3
#>     subject.summary subject.air_date subject.air_weekday
#>                                      subject.images.large
#>  1: http://lain.bgm.tv/pic/cover/l/9d/1f/313504_621SI.jpg
#>  2: http://lain.bgm.tv/pic/cover/l/c0/de/217573_trel5.jpg
#>  3: http://lain.bgm.tv/pic/cover/l/0d/05/267251_d6IUk.jpg
#>  4: http://lain.bgm.tv/pic/cover/l/fd/d3/282499_Xm8tG.jpg
#>  5: http://lain.bgm.tv/pic/cover/l/f6/f2/273843_YMdDZ.jpg
#>  6: http://lain.bgm.tv/pic/cover/l/28/e9/228870_G9dlD.jpg
#>  7: http://lain.bgm.tv/pic/cover/l/be/38/190930_HSsyO.jpg
#>  8: http://lain.bgm.tv/pic/cover/l/97/d9/227600_7IwSI.jpg
#>  9: http://lain.bgm.tv/pic/cover/l/b7/b7/271724_32Jou.jpg
#> 10: http://lain.bgm.tv/pic/cover/l/42/71/272458_6600X.jpg
#> 11: http://lain.bgm.tv/pic/cover/l/68/b8/274613_sx1iX.jpg
#> 12: http://lain.bgm.tv/pic/cover/l/e1/bb/123006_8YW00.jpg
#> 13: http://lain.bgm.tv/pic/cover/l/9c/28/209963_w1pVz.jpg
#> 14: http://lain.bgm.tv/pic/cover/l/1e/e4/266647_O6BAq.jpg
#> 15: http://lain.bgm.tv/pic/cover/l/11/95/245130_Tfzrz.jpg
#> 16: http://lain.bgm.tv/pic/cover/l/09/39/251578_wZCQu.jpg
#> 17: http://lain.bgm.tv/pic/cover/l/a1/46/210409_bX2U0.jpg
#> 18: http://lain.bgm.tv/pic/cover/l/c9/b9/236722_2abB1.jpg
#> 19: http://lain.bgm.tv/pic/cover/l/67/c8/235836_JOYZo.jpg
#> 20: http://lain.bgm.tv/pic/cover/l/ad/bd/231037_HN07n.jpg
#> 21: http://lain.bgm.tv/pic/cover/l/18/bc/207195_OOyqz.jpg
#> 22: http://lain.bgm.tv/pic/cover/l/3e/06/220722_3330w.jpg
#> 23: http://lain.bgm.tv/pic/cover/l/1e/e2/183878_Fef1o.jpg
#> 24:  http://lain.bgm.tv/pic/cover/l/a4/34/45336_ftJ2J.jpg
#> 25: http://lain.bgm.tv/pic/cover/l/71/ff/174142_HirNZ.jpg
#> 26: http://lain.bgm.tv/pic/cover/l/83/ff/192356_bgfkI.jpg
#> 27:    http://lain.bgm.tv/pic/cover/l/f2/fc/324_psuXk.jpg
#> 28: http://lain.bgm.tv/pic/cover/l/4e/37/189708_Vukq5.jpg
#> 29: http://lain.bgm.tv/pic/cover/l/a2/8e/128526_wZngS.jpg
#> 30:    http://lain.bgm.tv/pic/cover/l/d5/6b/893_gEo72.jpg
#> 31:      http://lain.bgm.tv/pic/cover/l/c9/f0/8_wK0z3.jpg
#> 32:  http://lain.bgm.tv/pic/cover/l/69/06/85316_aW8c9.jpg
#> 33: http://lain.bgm.tv/pic/cover/l/cb/78/140001_Ew1mo.jpg
#> 34: http://lain.bgm.tv/pic/cover/l/4c/8f/144087_E3I19.jpg
#> 35:    http://lain.bgm.tv/pic/cover/l/3d/8e/843_6wj9L.jpg
#> 36: http://lain.bgm.tv/pic/cover/l/9c/17/180945_0vv68.jpg
#> 37: http://lain.bgm.tv/pic/cover/l/7b/3a/168395_wg6qC.jpg
#> 38:  http://lain.bgm.tv/pic/cover/l/58/93/15877_5lXEu.jpg
#> 39: http://lain.bgm.tv/pic/cover/l/c0/fe/139408_yqY6B.jpg
#> 40:   http://lain.bgm.tv/pic/cover/l/8b/0d/1352_33Ma5.jpg
#> 41: http://lain.bgm.tv/pic/cover/l/1e/73/136376_WywYA.jpg
#> 42: http://lain.bgm.tv/pic/cover/l/56/ca/126014_j9eTa.jpg
#> 43:   http://lain.bgm.tv/pic/cover/l/80/73/5337_0saWw.jpg
#> 44: http://lain.bgm.tv/pic/cover/l/0e/76/139201_BI8j0.jpg
#> 45: http://lain.bgm.tv/pic/cover/l/21/0a/127132_IBsnZ.jpg
#> 46: http://lain.bgm.tv/pic/cover/l/dc/3f/119224_X15XN.jpg
#> 47: http://lain.bgm.tv/pic/cover/l/3c/4d/110248_n1OoO.jpg
#> 48:  http://lain.bgm.tv/pic/cover/l/d5/d7/72941_q1DiW.jpg
#> 49: http://lain.bgm.tv/pic/cover/l/00/b9/104521_6m21b.jpg
#> 50:   http://lain.bgm.tv/pic/cover/l/53/6a/1453_iZIOZ.jpg
#>                                      subject.images.large
#>                                     subject.images.common
#>  1: http://lain.bgm.tv/pic/cover/c/9d/1f/313504_621SI.jpg
#>  2: http://lain.bgm.tv/pic/cover/c/c0/de/217573_trel5.jpg
#>  3: http://lain.bgm.tv/pic/cover/c/0d/05/267251_d6IUk.jpg
#>  4: http://lain.bgm.tv/pic/cover/c/fd/d3/282499_Xm8tG.jpg
#>  5: http://lain.bgm.tv/pic/cover/c/f6/f2/273843_YMdDZ.jpg
#>  6: http://lain.bgm.tv/pic/cover/c/28/e9/228870_G9dlD.jpg
#>  7: http://lain.bgm.tv/pic/cover/c/be/38/190930_HSsyO.jpg
#>  8: http://lain.bgm.tv/pic/cover/c/97/d9/227600_7IwSI.jpg
#>  9: http://lain.bgm.tv/pic/cover/c/b7/b7/271724_32Jou.jpg
#> 10: http://lain.bgm.tv/pic/cover/c/42/71/272458_6600X.jpg
#> 11: http://lain.bgm.tv/pic/cover/c/68/b8/274613_sx1iX.jpg
#> 12: http://lain.bgm.tv/pic/cover/c/e1/bb/123006_8YW00.jpg
#> 13: http://lain.bgm.tv/pic/cover/c/9c/28/209963_w1pVz.jpg
#> 14: http://lain.bgm.tv/pic/cover/c/1e/e4/266647_O6BAq.jpg
#> 15: http://lain.bgm.tv/pic/cover/c/11/95/245130_Tfzrz.jpg
#> 16: http://lain.bgm.tv/pic/cover/c/09/39/251578_wZCQu.jpg
#> 17: http://lain.bgm.tv/pic/cover/c/a1/46/210409_bX2U0.jpg
#> 18: http://lain.bgm.tv/pic/cover/c/c9/b9/236722_2abB1.jpg
#> 19: http://lain.bgm.tv/pic/cover/c/67/c8/235836_JOYZo.jpg
#> 20: http://lain.bgm.tv/pic/cover/c/ad/bd/231037_HN07n.jpg
#> 21: http://lain.bgm.tv/pic/cover/c/18/bc/207195_OOyqz.jpg
#> 22: http://lain.bgm.tv/pic/cover/c/3e/06/220722_3330w.jpg
#> 23: http://lain.bgm.tv/pic/cover/c/1e/e2/183878_Fef1o.jpg
#> 24:  http://lain.bgm.tv/pic/cover/c/a4/34/45336_ftJ2J.jpg
#> 25: http://lain.bgm.tv/pic/cover/c/71/ff/174142_HirNZ.jpg
#> 26: http://lain.bgm.tv/pic/cover/c/83/ff/192356_bgfkI.jpg
#> 27:    http://lain.bgm.tv/pic/cover/c/f2/fc/324_psuXk.jpg
#> 28: http://lain.bgm.tv/pic/cover/c/4e/37/189708_Vukq5.jpg
#> 29: http://lain.bgm.tv/pic/cover/c/a2/8e/128526_wZngS.jpg
#> 30:    http://lain.bgm.tv/pic/cover/c/d5/6b/893_gEo72.jpg
#> 31:      http://lain.bgm.tv/pic/cover/c/c9/f0/8_wK0z3.jpg
#> 32:  http://lain.bgm.tv/pic/cover/c/69/06/85316_aW8c9.jpg
#> 33: http://lain.bgm.tv/pic/cover/c/cb/78/140001_Ew1mo.jpg
#> 34: http://lain.bgm.tv/pic/cover/c/4c/8f/144087_E3I19.jpg
#> 35:    http://lain.bgm.tv/pic/cover/c/3d/8e/843_6wj9L.jpg
#> 36: http://lain.bgm.tv/pic/cover/c/9c/17/180945_0vv68.jpg
#> 37: http://lain.bgm.tv/pic/cover/c/7b/3a/168395_wg6qC.jpg
#> 38:  http://lain.bgm.tv/pic/cover/c/58/93/15877_5lXEu.jpg
#> 39: http://lain.bgm.tv/pic/cover/c/c0/fe/139408_yqY6B.jpg
#> 40:   http://lain.bgm.tv/pic/cover/c/8b/0d/1352_33Ma5.jpg
#> 41: http://lain.bgm.tv/pic/cover/c/1e/73/136376_WywYA.jpg
#> 42: http://lain.bgm.tv/pic/cover/c/56/ca/126014_j9eTa.jpg
#> 43:   http://lain.bgm.tv/pic/cover/c/80/73/5337_0saWw.jpg
#> 44: http://lain.bgm.tv/pic/cover/c/0e/76/139201_BI8j0.jpg
#> 45: http://lain.bgm.tv/pic/cover/c/21/0a/127132_IBsnZ.jpg
#> 46: http://lain.bgm.tv/pic/cover/c/dc/3f/119224_X15XN.jpg
#> 47: http://lain.bgm.tv/pic/cover/c/3c/4d/110248_n1OoO.jpg
#> 48:  http://lain.bgm.tv/pic/cover/c/d5/d7/72941_q1DiW.jpg
#> 49: http://lain.bgm.tv/pic/cover/c/00/b9/104521_6m21b.jpg
#> 50:   http://lain.bgm.tv/pic/cover/c/53/6a/1453_iZIOZ.jpg
#>                                     subject.images.common
#>                                     subject.images.medium
#>  1: http://lain.bgm.tv/pic/cover/m/9d/1f/313504_621SI.jpg
#>  2: http://lain.bgm.tv/pic/cover/m/c0/de/217573_trel5.jpg
#>  3: http://lain.bgm.tv/pic/cover/m/0d/05/267251_d6IUk.jpg
#>  4: http://lain.bgm.tv/pic/cover/m/fd/d3/282499_Xm8tG.jpg
#>  5: http://lain.bgm.tv/pic/cover/m/f6/f2/273843_YMdDZ.jpg
#>  6: http://lain.bgm.tv/pic/cover/m/28/e9/228870_G9dlD.jpg
#>  7: http://lain.bgm.tv/pic/cover/m/be/38/190930_HSsyO.jpg
#>  8: http://lain.bgm.tv/pic/cover/m/97/d9/227600_7IwSI.jpg
#>  9: http://lain.bgm.tv/pic/cover/m/b7/b7/271724_32Jou.jpg
#> 10: http://lain.bgm.tv/pic/cover/m/42/71/272458_6600X.jpg
#> 11: http://lain.bgm.tv/pic/cover/m/68/b8/274613_sx1iX.jpg
#> 12: http://lain.bgm.tv/pic/cover/m/e1/bb/123006_8YW00.jpg
#> 13: http://lain.bgm.tv/pic/cover/m/9c/28/209963_w1pVz.jpg
#> 14: http://lain.bgm.tv/pic/cover/m/1e/e4/266647_O6BAq.jpg
#> 15: http://lain.bgm.tv/pic/cover/m/11/95/245130_Tfzrz.jpg
#> 16: http://lain.bgm.tv/pic/cover/m/09/39/251578_wZCQu.jpg
#> 17: http://lain.bgm.tv/pic/cover/m/a1/46/210409_bX2U0.jpg
#> 18: http://lain.bgm.tv/pic/cover/m/c9/b9/236722_2abB1.jpg
#> 19: http://lain.bgm.tv/pic/cover/m/67/c8/235836_JOYZo.jpg
#> 20: http://lain.bgm.tv/pic/cover/m/ad/bd/231037_HN07n.jpg
#> 21: http://lain.bgm.tv/pic/cover/m/18/bc/207195_OOyqz.jpg
#> 22: http://lain.bgm.tv/pic/cover/m/3e/06/220722_3330w.jpg
#> 23: http://lain.bgm.tv/pic/cover/m/1e/e2/183878_Fef1o.jpg
#> 24:  http://lain.bgm.tv/pic/cover/m/a4/34/45336_ftJ2J.jpg
#> 25: http://lain.bgm.tv/pic/cover/m/71/ff/174142_HirNZ.jpg
#> 26: http://lain.bgm.tv/pic/cover/m/83/ff/192356_bgfkI.jpg
#> 27:    http://lain.bgm.tv/pic/cover/m/f2/fc/324_psuXk.jpg
#> 28: http://lain.bgm.tv/pic/cover/m/4e/37/189708_Vukq5.jpg
#> 29: http://lain.bgm.tv/pic/cover/m/a2/8e/128526_wZngS.jpg
#> 30:    http://lain.bgm.tv/pic/cover/m/d5/6b/893_gEo72.jpg
#> 31:      http://lain.bgm.tv/pic/cover/m/c9/f0/8_wK0z3.jpg
#> 32:  http://lain.bgm.tv/pic/cover/m/69/06/85316_aW8c9.jpg
#> 33: http://lain.bgm.tv/pic/cover/m/cb/78/140001_Ew1mo.jpg
#> 34: http://lain.bgm.tv/pic/cover/m/4c/8f/144087_E3I19.jpg
#> 35:    http://lain.bgm.tv/pic/cover/m/3d/8e/843_6wj9L.jpg
#> 36: http://lain.bgm.tv/pic/cover/m/9c/17/180945_0vv68.jpg
#> 37: http://lain.bgm.tv/pic/cover/m/7b/3a/168395_wg6qC.jpg
#> 38:  http://lain.bgm.tv/pic/cover/m/58/93/15877_5lXEu.jpg
#> 39: http://lain.bgm.tv/pic/cover/m/c0/fe/139408_yqY6B.jpg
#> 40:   http://lain.bgm.tv/pic/cover/m/8b/0d/1352_33Ma5.jpg
#> 41: http://lain.bgm.tv/pic/cover/m/1e/73/136376_WywYA.jpg
#> 42: http://lain.bgm.tv/pic/cover/m/56/ca/126014_j9eTa.jpg
#> 43:   http://lain.bgm.tv/pic/cover/m/80/73/5337_0saWw.jpg
#> 44: http://lain.bgm.tv/pic/cover/m/0e/76/139201_BI8j0.jpg
#> 45: http://lain.bgm.tv/pic/cover/m/21/0a/127132_IBsnZ.jpg
#> 46: http://lain.bgm.tv/pic/cover/m/dc/3f/119224_X15XN.jpg
#> 47: http://lain.bgm.tv/pic/cover/m/3c/4d/110248_n1OoO.jpg
#> 48:  http://lain.bgm.tv/pic/cover/m/d5/d7/72941_q1DiW.jpg
#> 49: http://lain.bgm.tv/pic/cover/m/00/b9/104521_6m21b.jpg
#> 50:   http://lain.bgm.tv/pic/cover/m/53/6a/1453_iZIOZ.jpg
#>                                     subject.images.medium
#>                                      subject.images.small
#>  1: http://lain.bgm.tv/pic/cover/s/9d/1f/313504_621SI.jpg
#>  2: http://lain.bgm.tv/pic/cover/s/c0/de/217573_trel5.jpg
#>  3: http://lain.bgm.tv/pic/cover/s/0d/05/267251_d6IUk.jpg
#>  4: http://lain.bgm.tv/pic/cover/s/fd/d3/282499_Xm8tG.jpg
#>  5: http://lain.bgm.tv/pic/cover/s/f6/f2/273843_YMdDZ.jpg
#>  6: http://lain.bgm.tv/pic/cover/s/28/e9/228870_G9dlD.jpg
#>  7: http://lain.bgm.tv/pic/cover/s/be/38/190930_HSsyO.jpg
#>  8: http://lain.bgm.tv/pic/cover/s/97/d9/227600_7IwSI.jpg
#>  9: http://lain.bgm.tv/pic/cover/s/b7/b7/271724_32Jou.jpg
#> 10: http://lain.bgm.tv/pic/cover/s/42/71/272458_6600X.jpg
#> 11: http://lain.bgm.tv/pic/cover/s/68/b8/274613_sx1iX.jpg
#> 12: http://lain.bgm.tv/pic/cover/s/e1/bb/123006_8YW00.jpg
#> 13: http://lain.bgm.tv/pic/cover/s/9c/28/209963_w1pVz.jpg
#> 14: http://lain.bgm.tv/pic/cover/s/1e/e4/266647_O6BAq.jpg
#> 15: http://lain.bgm.tv/pic/cover/s/11/95/245130_Tfzrz.jpg
#> 16: http://lain.bgm.tv/pic/cover/s/09/39/251578_wZCQu.jpg
#> 17: http://lain.bgm.tv/pic/cover/s/a1/46/210409_bX2U0.jpg
#> 18: http://lain.bgm.tv/pic/cover/s/c9/b9/236722_2abB1.jpg
#> 19: http://lain.bgm.tv/pic/cover/s/67/c8/235836_JOYZo.jpg
#> 20: http://lain.bgm.tv/pic/cover/s/ad/bd/231037_HN07n.jpg
#> 21: http://lain.bgm.tv/pic/cover/s/18/bc/207195_OOyqz.jpg
#> 22: http://lain.bgm.tv/pic/cover/s/3e/06/220722_3330w.jpg
#> 23: http://lain.bgm.tv/pic/cover/s/1e/e2/183878_Fef1o.jpg
#> 24:  http://lain.bgm.tv/pic/cover/s/a4/34/45336_ftJ2J.jpg
#> 25: http://lain.bgm.tv/pic/cover/s/71/ff/174142_HirNZ.jpg
#> 26: http://lain.bgm.tv/pic/cover/s/83/ff/192356_bgfkI.jpg
#> 27:    http://lain.bgm.tv/pic/cover/s/f2/fc/324_psuXk.jpg
#> 28: http://lain.bgm.tv/pic/cover/s/4e/37/189708_Vukq5.jpg
#> 29: http://lain.bgm.tv/pic/cover/s/a2/8e/128526_wZngS.jpg
#> 30:    http://lain.bgm.tv/pic/cover/s/d5/6b/893_gEo72.jpg
#> 31:      http://lain.bgm.tv/pic/cover/s/c9/f0/8_wK0z3.jpg
#> 32:  http://lain.bgm.tv/pic/cover/s/69/06/85316_aW8c9.jpg
#> 33: http://lain.bgm.tv/pic/cover/s/cb/78/140001_Ew1mo.jpg
#> 34: http://lain.bgm.tv/pic/cover/s/4c/8f/144087_E3I19.jpg
#> 35:    http://lain.bgm.tv/pic/cover/s/3d/8e/843_6wj9L.jpg
#> 36: http://lain.bgm.tv/pic/cover/s/9c/17/180945_0vv68.jpg
#> 37: http://lain.bgm.tv/pic/cover/s/7b/3a/168395_wg6qC.jpg
#> 38:  http://lain.bgm.tv/pic/cover/s/58/93/15877_5lXEu.jpg
#> 39: http://lain.bgm.tv/pic/cover/s/c0/fe/139408_yqY6B.jpg
#> 40:   http://lain.bgm.tv/pic/cover/s/8b/0d/1352_33Ma5.jpg
#> 41: http://lain.bgm.tv/pic/cover/s/1e/73/136376_WywYA.jpg
#> 42: http://lain.bgm.tv/pic/cover/s/56/ca/126014_j9eTa.jpg
#> 43:   http://lain.bgm.tv/pic/cover/s/80/73/5337_0saWw.jpg
#> 44: http://lain.bgm.tv/pic/cover/s/0e/76/139201_BI8j0.jpg
#> 45: http://lain.bgm.tv/pic/cover/s/21/0a/127132_IBsnZ.jpg
#> 46: http://lain.bgm.tv/pic/cover/s/dc/3f/119224_X15XN.jpg
#> 47: http://lain.bgm.tv/pic/cover/s/3c/4d/110248_n1OoO.jpg
#> 48:  http://lain.bgm.tv/pic/cover/s/d5/d7/72941_q1DiW.jpg
#> 49: http://lain.bgm.tv/pic/cover/s/00/b9/104521_6m21b.jpg
#> 50:   http://lain.bgm.tv/pic/cover/s/53/6a/1453_iZIOZ.jpg
#>                                      subject.images.small
#>                                       subject.images.grid
#>  1: http://lain.bgm.tv/pic/cover/g/9d/1f/313504_621SI.jpg
#>  2: http://lain.bgm.tv/pic/cover/g/c0/de/217573_trel5.jpg
#>  3: http://lain.bgm.tv/pic/cover/g/0d/05/267251_d6IUk.jpg
#>  4: http://lain.bgm.tv/pic/cover/g/fd/d3/282499_Xm8tG.jpg
#>  5: http://lain.bgm.tv/pic/cover/g/f6/f2/273843_YMdDZ.jpg
#>  6: http://lain.bgm.tv/pic/cover/g/28/e9/228870_G9dlD.jpg
#>  7: http://lain.bgm.tv/pic/cover/g/be/38/190930_HSsyO.jpg
#>  8: http://lain.bgm.tv/pic/cover/g/97/d9/227600_7IwSI.jpg
#>  9: http://lain.bgm.tv/pic/cover/g/b7/b7/271724_32Jou.jpg
#> 10: http://lain.bgm.tv/pic/cover/g/42/71/272458_6600X.jpg
#> 11: http://lain.bgm.tv/pic/cover/g/68/b8/274613_sx1iX.jpg
#> 12: http://lain.bgm.tv/pic/cover/g/e1/bb/123006_8YW00.jpg
#> 13: http://lain.bgm.tv/pic/cover/g/9c/28/209963_w1pVz.jpg
#> 14: http://lain.bgm.tv/pic/cover/g/1e/e4/266647_O6BAq.jpg
#> 15: http://lain.bgm.tv/pic/cover/g/11/95/245130_Tfzrz.jpg
#> 16: http://lain.bgm.tv/pic/cover/g/09/39/251578_wZCQu.jpg
#> 17: http://lain.bgm.tv/pic/cover/g/a1/46/210409_bX2U0.jpg
#> 18: http://lain.bgm.tv/pic/cover/g/c9/b9/236722_2abB1.jpg
#> 19: http://lain.bgm.tv/pic/cover/g/67/c8/235836_JOYZo.jpg
#> 20: http://lain.bgm.tv/pic/cover/g/ad/bd/231037_HN07n.jpg
#> 21: http://lain.bgm.tv/pic/cover/g/18/bc/207195_OOyqz.jpg
#> 22: http://lain.bgm.tv/pic/cover/g/3e/06/220722_3330w.jpg
#> 23: http://lain.bgm.tv/pic/cover/g/1e/e2/183878_Fef1o.jpg
#> 24:  http://lain.bgm.tv/pic/cover/g/a4/34/45336_ftJ2J.jpg
#> 25: http://lain.bgm.tv/pic/cover/g/71/ff/174142_HirNZ.jpg
#> 26: http://lain.bgm.tv/pic/cover/g/83/ff/192356_bgfkI.jpg
#> 27:    http://lain.bgm.tv/pic/cover/g/f2/fc/324_psuXk.jpg
#> 28: http://lain.bgm.tv/pic/cover/g/4e/37/189708_Vukq5.jpg
#> 29: http://lain.bgm.tv/pic/cover/g/a2/8e/128526_wZngS.jpg
#> 30:    http://lain.bgm.tv/pic/cover/g/d5/6b/893_gEo72.jpg
#> 31:      http://lain.bgm.tv/pic/cover/g/c9/f0/8_wK0z3.jpg
#> 32:  http://lain.bgm.tv/pic/cover/g/69/06/85316_aW8c9.jpg
#> 33: http://lain.bgm.tv/pic/cover/g/cb/78/140001_Ew1mo.jpg
#> 34: http://lain.bgm.tv/pic/cover/g/4c/8f/144087_E3I19.jpg
#> 35:    http://lain.bgm.tv/pic/cover/g/3d/8e/843_6wj9L.jpg
#> 36: http://lain.bgm.tv/pic/cover/g/9c/17/180945_0vv68.jpg
#> 37: http://lain.bgm.tv/pic/cover/g/7b/3a/168395_wg6qC.jpg
#> 38:  http://lain.bgm.tv/pic/cover/g/58/93/15877_5lXEu.jpg
#> 39: http://lain.bgm.tv/pic/cover/g/c0/fe/139408_yqY6B.jpg
#> 40:   http://lain.bgm.tv/pic/cover/g/8b/0d/1352_33Ma5.jpg
#> 41: http://lain.bgm.tv/pic/cover/g/1e/73/136376_WywYA.jpg
#> 42: http://lain.bgm.tv/pic/cover/g/56/ca/126014_j9eTa.jpg
#> 43:   http://lain.bgm.tv/pic/cover/g/80/73/5337_0saWw.jpg
#> 44: http://lain.bgm.tv/pic/cover/g/0e/76/139201_BI8j0.jpg
#> 45: http://lain.bgm.tv/pic/cover/g/21/0a/127132_IBsnZ.jpg
#> 46: http://lain.bgm.tv/pic/cover/g/dc/3f/119224_X15XN.jpg
#> 47: http://lain.bgm.tv/pic/cover/g/3c/4d/110248_n1OoO.jpg
#> 48:  http://lain.bgm.tv/pic/cover/g/d5/d7/72941_q1DiW.jpg
#> 49: http://lain.bgm.tv/pic/cover/g/00/b9/104521_6m21b.jpg
#> 50:   http://lain.bgm.tv/pic/cover/g/53/6a/1453_iZIOZ.jpg
#>                                       subject.images.grid
#>     subject.collection.doing subject.eps subject.eps_count
#>  1:                       40        <NA>              <NA>
#>  2:                       21          13                13
#>  3:                        4           8                 8
#>  4:                        9           9                 9
#>  5:                      761          12                12
#>  6:                        9           8                 8
#>  7:                        8           6                 6
#>  8:                        9          10                10
#>  9:                      446          12                12
#> 10:                       21          10                10
#> 11:                      218          18                18
#> 12:                       21          10                10
#> 13:                       17         156               156
#> 14:                       35          10                10
#> 15:                       12          10                10
#> 16:                        1           9                 9
#> 17:                        7           5                 5
#> 18:                       18          10                10
#> 19:                        6          10                10
#> 20:                       10          10                10
#> 21:                      789          12                12
#> 22:                       10          10                10
#> 23:                     1423          14                14
#> 24:                       45          20                20
#> 25:                      296          11                11
#> 26:                       36          10                10
#> 27:                      580          26                26
#> 28:                       35          10                10
#> 29:                       13         156               156
#> 30:                      120          13                13
#> 31:                      122          25                25
#> 32:                       18          52                52
#> 33:                      998          25                25
#> 34:                       16          12                12
#> 35:                      215          13                13
#> 36:                        2          10                10
#> 37:                      127          13                13
#> 38:                      144          10                10
#> 39:                      419          25                25
#> 40:                       34          11                11
#> 41:                       10          12                12
#> 42:                       19          10                10
#> 43:                       48           7                 7
#> 44:                       24           8                 8
#> 45:                       50          10                10
#> 46:                       14          13                13
#> 47:                       58          10                10
#> 48:                     1126          25                25
#> 49:                       11          43                43
#> 50:                      532          39                39
#>     subject.collection.doing subject.eps subject.eps_count