Get bangumi user information

Get bangumi user information

bgm_user_info(username, format = "list", force = FALSE)

Arguments

username
A vector of user name or id.

format
Result will be a list of json data when set to l or list (default), a list of response() object when set to d or default, and a data.frame when set to t or table.

force
By default, you can get information of up to 10 users at a time, but if you want to break the limit, set it to TRUE but it's not recommended.

Value

A list of json data in list, response() object, or a data.frame contains all users' information.

Examples

library(bangumir)
bgm_user_info(username = "sai")

#> [[1]]
#> [[1]]$id
#> [1] 1
#> 
#> [[1]]$url
#> [1] "http://bgm.tv/user/sai"
#> 
#> [[1]]$username
#> [1] "sai"
#> 
#> [[1]]$nickname
#> [1] "Sai\U0001f596"
#> 
#> [[1]]$avatar
#> [[1]]$avatar$large
#> [1] "http://lain.bgm.tv/pic/user/l/000/00/00/1.jpg?r=1391790456"
#> 
#> [[1]]$avatar$medium
#> [1] "http://lain.bgm.tv/pic/user/m/000/00/00/1.jpg?r=1391790456"
#> 
#> [[1]]$avatar$small
#> [1] "http://lain.bgm.tv/pic/user/s/000/00/00/1.jpg?r=1391790456"
#> 
#> 
#> [[1]]$sign
#> [1] "Awesome!"
#> 
#> [[1]]$usergroup
#> [1] 1
#> 
#> 

bgm_user_info(username = c("sai", "infinityloop"), format = "t")

#>       id                             url     username      nickname
#> 1:     1          http://bgm.tv/user/sai          sai Sai\U0001f596
#> 2: 68998 http://bgm.tv/user/infinityloop infinityloop  InfinityLoop
#>                                                      avatar.large
#> 1:     http://lain.bgm.tv/pic/user/l/000/00/00/1.jpg?r=1391790456
#> 2: http://lain.bgm.tv/pic/user/l/000/06/89/68998.jpg?r=1607235401
#>                                                     avatar.medium
#> 1:     http://lain.bgm.tv/pic/user/m/000/00/00/1.jpg?r=1391790456
#> 2: http://lain.bgm.tv/pic/user/m/000/06/89/68998.jpg?r=1607235401
#>                                                      avatar.small       sign
#> 1:     http://lain.bgm.tv/pic/user/s/000/00/00/1.jpg?r=1391790456   Awesome!
#> 2: http://lain.bgm.tv/pic/user/s/000/06/89/68998.jpg?r=1607235401 深度隐藏宅
#>    usergroup
#> 1:         1
#> 2:        10