Get bangumi calendar

Get bangumi calendar

bgm_calendar(format = c("list", "table", "default"))

Arguments

format
Result will be a list of json data when set to list (default), response() object when set to default. Set to table will convert the episode list to a data.frame for ease of use.

Value

A list of response content or a response() object according to the format option.

Examples

result <- bgm_calendar("table")

# Bangumi list on Monday
result$mon

#>         id                          url type                        name
#>  1: 279059 http://bgm.tv/subject/279059    2    おちこぼれフルーツタルト
#>  2: 285901 http://bgm.tv/subject/285901    2      ゴールデンカムイ 第3期
#>  3: 289131 http://bgm.tv/subject/289131    2 A3! SEASON AUTUMN ＆ WINTER
#>  4: 290426 http://bgm.tv/subject/290426    2            魔王城でおやすみ
#>  5: 301233 http://bgm.tv/subject/301233    2     One Room サードシーズン
#>  6: 309910 http://bgm.tv/subject/309910    2            おそ松さん 第3期
#>  7: 315039 http://bgm.tv/subject/315039    2                    宿身小人
#>  8: 315133 http://bgm.tv/subject/315133    2            それだけがネック
#>  9: 316183 http://bgm.tv/subject/316183    2             梦幻书院 第四季
#> 10: 317276 http://bgm.tv/subject/317276    2                この空の下で
#> 11: 317653 http://bgm.tv/subject/317653    2                十年分の私へ
#> 12: 320427 http://bgm.tv/subject/320427    2        开心超人联盟之星之力
#>                   name_cn summary   air_date air_weekday rating.total
#>  1:          满溢的水果挞         2020-10-12           1          170
#>  2:       黄金神威 第三季         2020-10-05           1          126
#>  3: A3!满开剧团 秋&amp;冬         2020-10-12           1            5
#>  4:        在魔王城说晚安         2020-10-05           1          398
#>  5:       One Room 第三季         2020-10-05           1          207
#>  6:           阿松 第三季         2020-10-12           1           62
#>  7:              宿身小人         2020-10-05           1            2
#>  8:                               2020-10-12           1         <NA>
#>  9:       梦幻书院 第四季         2020-10-12           1            1
#> 10:                               2020-10-19           1            3
#> 11:                               2020-10-26           1           58
#> 12:                               2020-10-12           1         <NA>
#>     rating.count.10 rating.count.9 rating.count.8 rating.count.7 rating.count.6
#>  1:               6              6             24             56             51
#>  2:               5             21             66             26              5
#>  3:               0              0              2              2              1
#>  4:              15             11             70            170             94
#>  5:               7              4             55             94             38
#>  6:               2              2             20             30              5
#>  7:               0              0              0              1              1
#>  8:            <NA>           <NA>           <NA>           <NA>           <NA>
#>  9:               0              0              0              0              1
#> 10:               0              0              0              0              1
#> 11:               0              4              8             16             21
#> 12:            <NA>           <NA>           <NA>           <NA>           <NA>
#>     rating.count.5 rating.count.4 rating.count.3 rating.count.2 rating.count.1
#>  1:             18              6              2              1              0
#>  2:              2              1              0              0              0
#>  3:              0              0              0              0              0
#>  4:             27              4              3              3              1
#>  5:              7              1              1              0              0
#>  6:              3              0              0              0              0
#>  7:              0              0              0              0              0
#>  8:           <NA>           <NA>           <NA>           <NA>           <NA>
#>  9:              0              0              0              0              0
#> 10:              2              0              0              0              0
#> 11:              9              0              0              0              0
#> 12:           <NA>           <NA>           <NA>           <NA>           <NA>
#>     rating.score rank                                          images.large
#>  1:          6.6 3189 http://lain.bgm.tv/pic/cover/l/0d/2e/279059_NNOnk.jpg
#>  2:          7.9  482 http://lain.bgm.tv/pic/cover/l/72/f9/285901_94tQr.jpg
#>  3:          7.2 <NA> http://lain.bgm.tv/pic/cover/l/8a/f4/289131_wg9Cd.jpg
#>  4:          6.9 2615 http://lain.bgm.tv/pic/cover/l/b1/c3/290426_z4GRp.jpg
#>  5:          7.1 1813 http://lain.bgm.tv/pic/cover/l/40/5b/301233_mq8zr.jpg
#>  6:          7.3 1430 http://lain.bgm.tv/pic/cover/l/49/41/309910_jm70D.jpg
#>  7:          6.5 <NA> http://lain.bgm.tv/pic/cover/l/da/bc/315039_rRr01.jpg
#>  8:         <NA> <NA> http://lain.bgm.tv/pic/cover/l/20/a9/315133_1tyTo.jpg
#>  9:            6 <NA> http://lain.bgm.tv/pic/cover/l/2f/64/316183_1C1D7.jpg
#> 10:          5.3 <NA> http://lain.bgm.tv/pic/cover/l/2a/be/317276_YVLrl.jpg
#> 11:          6.6 2703 http://lain.bgm.tv/pic/cover/l/ab/9a/317653_mHcgo.jpg
#> 12:         <NA> <NA> http://lain.bgm.tv/pic/cover/l/76/ff/320427_384xY.jpg
#>                                             images.common
#>  1: http://lain.bgm.tv/pic/cover/c/0d/2e/279059_NNOnk.jpg
#>  2: http://lain.bgm.tv/pic/cover/c/72/f9/285901_94tQr.jpg
#>  3: http://lain.bgm.tv/pic/cover/c/8a/f4/289131_wg9Cd.jpg
#>  4: http://lain.bgm.tv/pic/cover/c/b1/c3/290426_z4GRp.jpg
#>  5: http://lain.bgm.tv/pic/cover/c/40/5b/301233_mq8zr.jpg
#>  6: http://lain.bgm.tv/pic/cover/c/49/41/309910_jm70D.jpg
#>  7: http://lain.bgm.tv/pic/cover/c/da/bc/315039_rRr01.jpg
#>  8: http://lain.bgm.tv/pic/cover/c/20/a9/315133_1tyTo.jpg
#>  9: http://lain.bgm.tv/pic/cover/c/2f/64/316183_1C1D7.jpg
#> 10: http://lain.bgm.tv/pic/cover/c/2a/be/317276_YVLrl.jpg
#> 11: http://lain.bgm.tv/pic/cover/c/ab/9a/317653_mHcgo.jpg
#> 12: http://lain.bgm.tv/pic/cover/c/76/ff/320427_384xY.jpg
#>                                             images.medium
#>  1: http://lain.bgm.tv/pic/cover/m/0d/2e/279059_NNOnk.jpg
#>  2: http://lain.bgm.tv/pic/cover/m/72/f9/285901_94tQr.jpg
#>  3: http://lain.bgm.tv/pic/cover/m/8a/f4/289131_wg9Cd.jpg
#>  4: http://lain.bgm.tv/pic/cover/m/b1/c3/290426_z4GRp.jpg
#>  5: http://lain.bgm.tv/pic/cover/m/40/5b/301233_mq8zr.jpg
#>  6: http://lain.bgm.tv/pic/cover/m/49/41/309910_jm70D.jpg
#>  7: http://lain.bgm.tv/pic/cover/m/da/bc/315039_rRr01.jpg
#>  8: http://lain.bgm.tv/pic/cover/m/20/a9/315133_1tyTo.jpg
#>  9: http://lain.bgm.tv/pic/cover/m/2f/64/316183_1C1D7.jpg
#> 10: http://lain.bgm.tv/pic/cover/m/2a/be/317276_YVLrl.jpg
#> 11: http://lain.bgm.tv/pic/cover/m/ab/9a/317653_mHcgo.jpg
#> 12: http://lain.bgm.tv/pic/cover/m/76/ff/320427_384xY.jpg
#>                                              images.small
#>  1: http://lain.bgm.tv/pic/cover/s/0d/2e/279059_NNOnk.jpg
#>  2: http://lain.bgm.tv/pic/cover/s/72/f9/285901_94tQr.jpg
#>  3: http://lain.bgm.tv/pic/cover/s/8a/f4/289131_wg9Cd.jpg
#>  4: http://lain.bgm.tv/pic/cover/s/b1/c3/290426_z4GRp.jpg
#>  5: http://lain.bgm.tv/pic/cover/s/40/5b/301233_mq8zr.jpg
#>  6: http://lain.bgm.tv/pic/cover/s/49/41/309910_jm70D.jpg
#>  7: http://lain.bgm.tv/pic/cover/s/da/bc/315039_rRr01.jpg
#>  8: http://lain.bgm.tv/pic/cover/s/20/a9/315133_1tyTo.jpg
#>  9: http://lain.bgm.tv/pic/cover/s/2f/64/316183_1C1D7.jpg
#> 10: http://lain.bgm.tv/pic/cover/s/2a/be/317276_YVLrl.jpg
#> 11: http://lain.bgm.tv/pic/cover/s/ab/9a/317653_mHcgo.jpg
#> 12: http://lain.bgm.tv/pic/cover/s/76/ff/320427_384xY.jpg
#>                                               images.grid collection.doing
#>  1: http://lain.bgm.tv/pic/cover/g/0d/2e/279059_NNOnk.jpg              738
#>  2: http://lain.bgm.tv/pic/cover/g/72/f9/285901_94tQr.jpg              468
#>  3: http://lain.bgm.tv/pic/cover/g/8a/f4/289131_wg9Cd.jpg               44
#>  4: http://lain.bgm.tv/pic/cover/g/b1/c3/290426_z4GRp.jpg             1316
#>  5: http://lain.bgm.tv/pic/cover/g/40/5b/301233_mq8zr.jpg              860
#>  6: http://lain.bgm.tv/pic/cover/g/49/41/309910_jm70D.jpg              352
#>  7: http://lain.bgm.tv/pic/cover/g/da/bc/315039_rRr01.jpg                5
#>  8: http://lain.bgm.tv/pic/cover/g/20/a9/315133_1tyTo.jpg                3
#>  9: http://lain.bgm.tv/pic/cover/g/2f/64/316183_1C1D7.jpg                1
#> 10: http://lain.bgm.tv/pic/cover/g/2a/be/317276_YVLrl.jpg             <NA>
#> 11: http://lain.bgm.tv/pic/cover/g/ab/9a/317653_mHcgo.jpg                3
#> 12: http://lain.bgm.tv/pic/cover/g/76/ff/320427_384xY.jpg             <NA>