Get summary of user's collection

Get summary of user's collection

bgm_latest(
  username,
  type = c("anime", "book", "game", "music", "real"),
  max = 25,
  format = c("list", "table", "default"),
  key = "bgm2725b1fa2fea5582"
)

Arguments

username
User name a user id.

type
Category type, anime, book, game, music and real are available.

max
Max result number, default is the 25 (Maximum value).

format
Result will be a list of json data when set to list (default), response() object when set to default, and list of data.table when set to table.

key
Your application key (ID). bangumir already comes with a key/secret but you are welcome to swap out with our own. Since these keys are shipped with the package, there is a small chance they could be voided if someone abuses the key. If you plan to use this in production, or for an internal tool, the recommended practice is to create a new application on Bangumi and use those keys for your purposes.

Value

A list of response content, response() object, or a data.table contains the latest collections.

Examples

result <- bgm_latest(username = "sai", format = "table")

#> Warning: Coercing LHS to a list

#> Warning: Coercing LHS to a list

#> Warning: Coercing LHS to a list

#> Warning: Coercing LHS to a list

#> Warning: Coercing LHS to a list

# Result of watching
result$do

#> [[1]]
#> [1] TRUE
#> 
#> [[2]]
#> [1] FALSE
#> 
#> [[3]]
#> [1] TRUE
#> 
#> [[4]]
#> [1] TRUE
#> 
#> [[5]]
#> [1] TRUE
#> 
#> [[6]]
#> [1] TRUE
#> 
#> [[7]]
#> [1] FALSE
#> 
#> [[8]]
#> [1] FALSE
#> 
#> [[9]]
#> [1] FALSE
#> 
#> [[10]]
#> [1] TRUE
#> 
#> [[11]]
#> [1] FALSE
#> 
#> [[12]]
#> [1] FALSE
#> 
#> [[13]]
#> [1] FALSE
#> 
#> [[14]]
#> [1] FALSE
#> 
#> $subject.type
#> [1] "anime"
#>